% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{add_item_to_rdf}
\alias{add_item_to_rdf}
\title{Title}
\usage{
add_item_to_rdf(
  object,
  subject,
  classes = NULL,
  dataProperties = NULL,
  objectProperties = NULL,
  namespaces = NULL
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
