% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{align_time_grid}
\alias{align_time_grid}
\title{The function aligns the frequency of the input time series with the output
frequency given as an argument using the specified aggregation function.}
\usage{
align_time_grid(
  data,
  measurementReadingType,
  outputTimeStep,
  aggregationFunction
)
}
\arguments{
\item{data}{<timeSeries> The time series that has to be aligned with an
output time step, i.e. with a specific frequency. If the
measurementReadingType of the series is not instantaneous, the data must
be converted first using the function clean_ts_integrate.}

\item{outputTimeStep}{<string> The frequency used to resample the input
time series for the alignment. It must be a string in ISO 8601 format
representing the time step (e.g. "15T","1H", "M", "D",...).}

\item{aggregationFunction}{<string> The aggregation function to use when}
}
\value{
data <timeSeries> The time series resampled with the specified
period and aggregation function.
}
\description{
The function aligns the frequency of the input time series with the output
frequency given as an argument using the specified aggregation function.
}
