% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{append_emissions_to_sensor}
\alias{append_emissions_to_sensor}
\title{Title}
\usage{
append_emissions_to_sensor(
  buildingsRdf,
  timeseriesObject,
  emissionsSubject,
  measuredProperty,
  frequency,
  energyTimeseriesSensor
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
