% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{calculate_indicator}
\alias{calculate_indicator}
\title{Title}
\usage{
calculate_indicator(
  data,
  indicator,
  consumptionColumn,
  baselineConsumptionColumn,
  energyPriceColumn,
  carbonEmissionsColumn,
  buildingGrossFloorArea,
  heatingDegreeDays18Column,
  coolingDegreeDays21Column
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
