% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{calculate_indicator_not_aggregable_by_time}
\alias{calculate_indicator_not_aggregable_by_time}
\title{Title}
\usage{
calculate_indicator_not_aggregable_by_time(
  indicator,
  annualEnergySavings,
  annualCostSavings,
  affectedBuildingArea,
  investment,
  discountRate,
  lifespan
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
