% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{calendar_components}
\alias{calendar_components}
\title{Time decomposition}
\usage{
calendar_components(data, localTimeZone = NULL, holidays = c(), inplace = T)
}
\arguments{
\item{data}{<data.frame> containing the series to transform. The time zone
of the datetimes must be UTC. The other variables describing the series are
directly bypassed to the output.}

\item{localTimeZone}{<string> specifying the local time zone related to the
building in analysis. The format of this time zones are defined by the IANA
Time Zone Database (https://www.iana.org/time-zones). This argument is
optional, by default no transformation to local time zone is done.}

\item{holidays}{<array> of dates classified as holidays}

\item{inplace}{<boolean> indicating if the output should be the original
data argument, plus the transformed objects -True- , or only the transformed
series -False.}
}
\value{
<data.frame> containing the same initial information of data
input argument, plus the calendar components as new columns.
}
\description{
Decompose the time in date, day of the year, day of the week, day of the
weekend, working day, non-working day, season, month, hour, minute, ...
}
