% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_if_complete.R
\name{checking_if_complete}
\alias{checking_if_complete}
\title{checking_if_complete}
\usage{
checking_if_complete(
  dns_names = NULL,
  username = "ubuntu",
  password = "password",
  unique_file = "complete",
  follow_file = NULL,
  keyfile = "/Users/fdrennan/fdren.pem"
)
}
\arguments{
\item{dns_names}{dns names}

\item{username}{username}

\item{password}{password}

\item{unique_file}{unique_file}

\item{keyfile}{keyfile}
}
\description{
Check to see if complete file exists
}
