% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{classification_dlc}
\alias{classification_dlc}
\title{Classify daily load curves based on the outputs of a clustering and a
new set of data}
\usage{
classification_dlc(
  data,
  consumptionFeature,
  outdoorTemperatureFeature,
  localTimeZone,
  clustering,
  methodPriority
)
}
\arguments{
\item{data}{<timeSeries> containing the time series for total energy consumption
of a building, the outdoor temperature, or whatever input is needed for clustering.}

\item{consumptionFeature}{<string> containing the column name the consumption feature 
in the data argument.}

\item{outdoorTemperatureFeature}{<string> containing the column name of the outdoor temperature feature
in the data argument.}

\item{localTimeZone}{<string> local time zone}

\item{clustering}{<object> clustering_dlc() output}

\item{methodPriority:}{<list of strings>. Possible values:
absoluteLoadCurvesCentroids: Based on the absolute consumption load
  curve
clusteringCentroids: Based on the inputs considered in the clustering
  procedure. Applying the same transformations done during that process.
classificationModel: Based on a classification model of the calendar
  features.}
}
\value{
dailyClassification: <timeSeries>
}
\description{
Classify daily load curves based on the outputs of a clustering and a
new set of data
}
