% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{clean_ts_integrate}
\alias{clean_ts_integrate}
\title{The function converts a cumulative (counter) or onChange (delta) measurement
to instantaneous.}
\usage{
clean_ts_integrate(data, measurementReadingType = "onChange")
}
\arguments{
\item{data}{<timeSeries> The cumulative or on-change time series that has to
be converted to instantaneous. An instantaneous measurement is a gauge
metric, in which the value can increase or decrease and measures a specific
instant in time.}

\item{measurementReadingType}{<enumeration> An argument used to define the
measurement reading type}
}
\value{
data <timeSeries> The cumulative or onChange time series with
the measurements converted to the instantaneous type.
}
\description{
The function converts a cumulative (counter) or onChange (delta) measurement
to instantaneous.
}
