% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{compute_device_aggregator_formula}
\alias{compute_device_aggregator_formula}
\title{Compute the specified formula of a device aggregator}
\usage{
compute_device_aggregator_formula(
  buildingsRdf,
  timeseriesObject,
  buildingSubject,
  formula,
  outputFrequency,
  aggFunctions,
  useEstimatedValues,
  ratioCorrection = F,
  minRatioCorrection = 0.7,
  transformToAggregatableMeasuredProperty = F,
  aggregatableMeasuredPropertyName = NULL,
  defaultFactorsByMeasuredProperty = NULL
)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{timeseriesObject}{<string> path of JSON files, or <list> of time series.}

\item{buildingSubject}{<uri> containing the building subject.}

\item{formula}{<string> describing the formula for the device aggregator. 
It consist on a sequence of arithmetical operations over one (or multiple) sensor 
identifier(s). The sensor identifiers must be written between prefix '<mi>' and suffix </mi>. 
On contrary, the operators are described between prefix <mo> and suffix </mo>.
Example of the sum of two sensors (let's identify them as 'ID1' and 'ID2'): 
'<mi>ID1</mi><mo>+</mo><mi>ID2</mi>'}

\item{outputFrequency}{<string> defining the frequency selected as output. 
It must follow ISO 8601 format representing the time step.}

\item{aggFunctions}{<string> describing the possible aggregation functions of the
resultant time series. Possible values: 'SUM', 'AVG', 'HDD', 'CDD'.}

\item{useEstimatedValues}{<boolean> describing if the estimated values of time series 
should be taken into account.}

\item{ratioCorrection}{<boolean> describing whether a ratio correction should be done, 
or not. Important to set to TRUE when time series contain gaps.}
}
\value{
<data.frame> containing the resultant time series.
}
\description{
This function obtains all the time series related with a device aggregator, aggregates them 
according to the device aggregator metadata and obtains a single resultant time series.
}
