% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_aws.R
\name{configure_aws}
\alias{configure_aws}
\title{configure_aws}
\usage{
configure_aws(
  aws_access_key_id = NA,
  aws_secret_access_key = NA,
  default.region = NA
)
}
\arguments{
\item{aws_access_key_id}{IAM access key ID}

\item{aws_secret_access_key}{IAM secret access key}

\item{default.region}{AWS preferred region}
}
\description{
configure_aws
}
