% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_get.R
\name{cost_get}
\alias{cost_get}
\title{cost_get
A function to get cost on a specified date range. It defaults to a monthly
range if from io are not specified}
\usage{
cost_get(from = NA, to = NA)
}
\arguments{
\item{from}{The earliest date to grab data. No longer than a year ago.}

\item{to}{The latest date to grab data. No later than the end of the month.}
}
\value{
Current AWS costs
}
\description{
cost_get
A function to get cost on a specified date range. It defaults to a monthly
range if from io are not specified
}
\examples{
cost_get(from = '2019-01-01', to = '2019-03-05')
}
