% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{data_transformation_wrapper}
\alias{data_transformation_wrapper}
\title{Description}
\usage{
data_transformation_wrapper(
  data,
  features,
  transformationSentences,
  param,
  transformationResults = list(),
  weatherDependenceByCluster = NULL,
  clusteringResults = NULL
)
}
\arguments{
\item{arg}{<>}
}
\description{
Description
}
