% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{decodeBinFromValue}
\alias{decodeBinFromValue}
\title{Encode value to binary representation}
\usage{
decodeBinFromValue(values, features)
}
\arguments{
\item{values}{<list> List of values to encode}

\item{features}{<list> List of features to encode
Feature description
feature = list(
   datatype = <type of data> (discrete | integer | float)
   min = <min value>,
   max = <max value>,
   nlevels = <number of levels>,
   levels = <levels> (optional: discrete type specific)
)}
}
\value{
<list> of encoded values
}
\description{
Encode value to binary representation
}
