% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{decodeValueFromBin}
\alias{decodeValueFromBin}
\title{Decode binary representation to value}
\usage{
decodeValueFromBin(binary, features)
}
\arguments{
\item{binary}{<integer> Binary encoded value}

\item{features}{<list> List of features to decode
Feature description
feature = list(
   datatype = <type of data> (discrete | integer | float)
   min = <min value>,
   max = <max value>,
   nlevels = <number of levels>,
   levels = <levels> (optional: discrete type specific)
)}
}
\value{
<list> of decoded values
}
\description{
Decode binary representation to value
}
