% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{degree_days}
\alias{degree_days}
\title{Calculate the degree-days with a desired output frequency and considering
cooling or heating mode.}
\usage{
degree_days(
  data,
  featuresName,
  localTimeZone,
  baseTemperature,
  outputFeaturesName = NULL,
  mode = "heating",
  outputTimeStep = "D"
)
}
\arguments{
\item{data}{<timeSeries> containing the series to transform.
Maximum input frequency is daily ("D") or higher ("H","15T",...).}

\item{localTimeZone}{<string> specifying the local time zone related to
the building in analysis. The format of this time zones are defined by
the IANA Time Zone Database (https://www.iana.org/time-zones). This
argument is optional, by default no transformation to local time zone is
done.}

\item{baseTemperature}{<float> describing the Balance Point
Temperature (BPT) used in the calculation. Below BPT in heating mode,
heat would be required by the building. The contrary in the case of
cooling, over BPT in cooling mode}

\item{mode}{<string> describing the calculation mode, which could be
"cooling" or "heating". By default, "heating" is configured.}

\item{outputTimeStep}{<string> The frequency used to resample the daily
degree days. It must be a string in ISO 8601 format representing the
time step. Only yearly ("Y"), monthly ("M"), daily ("D") output time
steps are allowed.}

\item{featuresNames}{<string> giving the column name of the outdoor temperature feature.}

\item{outputFeaturesNames}{<string> giving the column name used as output of the transformation.
By default, "heating" or "cooling" depending the mode used in the transformation.}
}
\value{
degreeDays <timeSeries> in the outputTimeStep of the heating or
cooling degree days.
}
\description{
Calculate the degree-days with a desired output frequency and considering
cooling or heating mode.
}
