% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{degree_raw}
\alias{degree_raw}
\title{Calculate the difference between outdoor temperature and a base temperature,
without considering the frequency of the original data.}
\usage{
degree_raw(
  data,
  featuresName,
  baseTemperature,
  outputFeaturesName = NULL,
  mode = "heating",
  inplace = T
)
}
\arguments{
\item{data}{<timeSeries> containing the series to transform.}

\item{baseTemperature}{<float> describing the Balance Point Temperature (BPT)
used in the calculation. Below BPT in heating mode, heat would be required by
the building. The contrary in the case of cooling, over BPT in cooling mode.}

\item{featuresNames}{<string> giving the column name of the outdoor temperature feature.}

\item{outputFeaturesNames}{<string> giving the column name used as output of the transformation.
By default, "heating" or "cooling" depending the mode used in the transformation.}

\item{mode:}{<string> describing the calculation mode, which could be "cooling"
or "heating". By default, "heating" is configured.}

\item{inplace:}{<boolean> indicating if the output should be the original data argument, 
plus the transformed objects -True- , or only the transformed series -False.}
}
\value{
<timeSeries< of the difference between the temperature argument
and the selected base temperature, mantaining the original frequency of
temperature.
}
\description{
Calculate the difference between outdoor temperature and a base temperature,
without considering the frequency of the original data.
}
