% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_disruptive_period}
\alias{detect_disruptive_period}
\title{Detect disruptive periods in consumption time series.}
\usage{
detect_disruptive_period(
  data,
  consumptionColumn,
  timeColumn,
  temperatureColumn,
  tz,
  minIniDate,
  maxIniDate,
  minEndDate,
  maxEndDate,
  checkFor = "decrement",
  minDecrementPercentualAffectation = 30,
  minIncrementPercentualAffectation = 60
)
}
\arguments{
\item{data}{<timeSeries> Time serie with potential anomalies in
values}

\item{timeColumn}{<string> Time column in data time series}

\item{temperatureColumn}{<string> Temperature column in data time series}

\item{tz}{<string> specifying the local time zone related to
the building in analysis. The format of this time zones are defined by
the IANA Time Zone Database (https://www.iana.org/time-zones).}

\item{minIniDate}{<date> Minimum date to start looking for anomalies}

\item{maxIniDate}{<date> Maximum date to start looking for anomalies}

\item{minEndDate}{<date> Minimum date to stop looking for anomalies}

\item{maxEndDate}{<date> Maximum date to stop looking for anomalies}

\item{checkFor}{<string> Which effect will the disruptive
event generate? e.g. the SARS-CoV2 pandemia decrease the energy consumption of tertiary buildings,
so in that case we should check for decrement in consumption 
Possible values: decrement, increment, incrementAndDecrement.}

\item{minDecrementPercentualAffectation}{<float> indicating the minimum
decrement of consumption (vs. baseline) to be considered as abnormal.
Default 30%.}

\item{minIncrementPercentualAffectation}{<float> indicating the minimum
increment of consumption (vs. baseline) to be considered as abnormal.
Default 60%.}

\item{consumptionValue}{<string> Consumption column in data time series}
}
\value{
<data.frame> with the period of time (min-max) with abnormal 
consumption.
}
\description{
The function detects periods of time with abnormal consumption.
It evaluates multiple time periods in order to find the best one
fitting a model which considers the relationship of consumption 
against weekday and temperature.
}
