% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_holidays_in_tertiary_buildings}
\alias{detect_holidays_in_tertiary_buildings}
\title{The function detects holidays period in tertiary building time serie.
Analysis of turnpoints in density distribution in order to identify
"over represented" low consumption dates. Additional postprocessing is done
in order to take care of additional constraints related to holidays period}
\usage{
detect_holidays_in_tertiary_buildings(
  data,
  consumptionColumn,
  timeColumn,
  plotDensity = T,
  ignoreDates = c(),
  tz = "UTC"
)
}
\arguments{
\item{data}{<timeSeries> Time serie with potential anomalies in
values}

\item{consumptionColumn}{<string> Value column}

\item{timeColumn}{<string> Time column}

\item{ignoreDates}{<list> Ignore list of dates in time serie}

\item{plotDensiry}{<boolean> Plot density function used to obtain
holidays consumption threshold}
}
\value{
List of dates with classified as holidays
}
\description{
The function detects holidays period in tertiary building time serie.
Analysis of turnpoints in density distribution in order to identify
"over represented" low consumption dates. Additional postprocessing is done
in order to take care of additional constraints related to holidays period
}
