% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_profiled_data}
\alias{detect_profiled_data}
\title{Detect profiled days in data}
\usage{
detect_profiled_data(data)
}
\arguments{
\item{data}{timeSeries Data time serie with potential profiled content}
}
\value{
list <date> List of dates with profiled data
}
\description{
Detect profiled days in data
}
