% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_static_min_max_outliers}
\alias{detect_static_min_max_outliers}
\title{Detect which numerical elements are outside the min-max range.}
\usage{
detect_static_min_max_outliers(
  data,
  min,
  max,
  includeMin = TRUE,
  includeMax = TRUE
)
}
\arguments{
\item{data}{<float/integer> list The numerical elements to be evaluated}

\item{min}{<float/integer> The minimum value of the range.}

\item{max}{<float/integer> The maximum value of the range.}

\item{includedMin}{<boolean> An optional argument setting if the
minimum value should be included as valid value (true), or not (false).
Default is true.}

\item{<includedMax>}{boolean An optional argument setting if the maximum
value should be included as valid value (true), or not (false).
Default is true.}
}
\description{
Detect which numerical elements are outside the min-max range.
}
