% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_time_step}
\alias{detect_time_step}
\title{The function infers, i.e. automatically deduce from the input data, the
minimum time step (frequency) that can be used for the input time series.}
\usage{
detect_time_step(data, maxMissingTimeSteps = 0, approxTimeSteps = FALSE)
}
\arguments{
\item{data}{<timeSeries> Input time series whose time step has to be detected.}

\item{maxMissingTimeSteps<int>}{Optional tolerance threshold specifying the
maximum tolerable percentage of missing values on the total length of the time
series.}
}
\value{
timeStep <string> A string in ISO 8601 format representing the time
step (e.g. "15T","1H", "3M", "1D" ,...). If no frequency can be detected,
the function will return None.
}
\description{
The function infers, i.e. automatically deduce from the input data, the
minimum time step (frequency) that can be used for the input time series.
}
