% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_ts_min_max_outliers}
\alias{detect_ts_min_max_outliers}
\title{Detect elements of the time series outside a minimum and maximum range.}
\usage{
detect_ts_min_max_outliers(data, min, max, minSeries = NULL, maxSeries = NULL)
}
\arguments{
\item{data}{<timeSeries> An argument containing the time series from which
the outliers need to be detected.}

\item{min:}{float The minimum value allowed for each element of the time
series.}

\item{max:}{float The maximum value allowed for each element of the time
series.}

\item{minSeries:}{timeSeries. An optional argument defining a time series
with minimum allowed values.}

\item{maxSeries:}{timeSeries. An optional argument defining a time series
with maximum allowed values.}
}
\value{
outliers <boolean> timeSeries object using the original time period
and frequency, only assigning True values when an element should be
considered as an outlier.
}
\description{
Additionally, with the minSeries and maxSeries arguments, this ranges can
be set differently along the period. When using this feature, the timestep
of both minSeries and maxSeries need to be resampled to the original
frequency of the data time series, applying forward fill if is needed
(Remember the time stamps of each time series element always represent the
begining of the time step).
}
