\name{E.coli_iAF1260}
\docType{data}
\alias{E.coli_iAF1260}
\title{Ecoli dataset with ORFs and thermodynamic information}
\description{
A genome-scale metabolic reconstruction for Escherichia coli K-12 MG1655 that accounts for 1260 ORFs and thermodynamic information
}
\usage{E.coli_iAF1260}
\format{An sbml object of class rsbml}
\source{Molecular Systems Biology, 3:121(2007)}
\references{
Feist, A.M., Henry, C.S., Reed, J.L., Krummenacker, M., Joyce, A.R., Karp, P.D., Broadbelt, L.J., Hatzimanikatis, V., Palsson, B.O., 
\emph{A genome-scale metabolic reconstruction for Escherichia coli K-12 MG1655 that accounts for 1260 ORFs and thermodynamic information}, olecular Systems Biology, 3:121 (2007)
}
\keyword{datasets}
