\name{E.coli_iJR904}
\encoding{latin1}
\docType{data}
\alias{E.coli_iJR904}
\title{Ecoli genome-scale model}
\description{
 An expanded genome-scale model of Escherichia coli K-12 (iJR904 GSM/GPR)
}
\usage{E.coli_iJR904}
\format{An sbml object of class rsbml}
\source{Genome Biology, 4(9): R54.1-R54.12 (2003)}
\references{
Reed, J.L., Vo, T.D., Schilling, C.H., and Palsson, B.O., 
\emph{An expanded genome-scale model of Escherichia coli K-12 (iJR904 GSM/GPR)}, Genome Biology, 4(9): R54.1-R54.12 (2003).
}
\keyword{ecoli2}
