% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_image_create.R
\name{ec2_image_create}
\alias{ec2_image_create}
\title{ec2_image_create}
\usage{
ec2_image_create(name = NA, instance_id = NA)
}
\arguments{
\item{name}{A name for the imange}

\item{instance_id}{And instance Id}
}
\description{
ec2_image_create
}
