% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_utilities.R
\name{ec2_instance_create}
\alias{ec2_instance_create}
\title{ec2_instance_create}
\usage{
ec2_instance_create(ImageId = NA, InstanceType = "t2.nano", min = 1,
  max = 1)
}
\arguments{
\item{ImageId}{An aws ec2 id: i.e., 'ami-0174e69c12bae5410'}

\item{InstanceType}{See \url{https://aws.amazon.com/ec2/instance-types/}}

\item{resource}{The result of resource_ec2}
}
\description{
ec2_instance_create
}
