% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_instance_modify.R
\name{ec2_instance_modify}
\alias{ec2_instance_modify}
\title{ec2_instance_modify}
\usage{
ec2_instance_modify(
  instance_id = NA,
  attribute = "instanceType",
  value = "t2.small"
)
}
\arguments{
\item{instance_id}{An aws ec2 id: i.e., 'i-034e6090b1eb879e7'}

\item{attribute}{'instanceType',}

\item{value}{'t2.small'}
}
\description{
ec2_instance_modify
}
