% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_instance_script.R
\name{ec2_instance_script}
\alias{ec2_instance_script}
\title{ec2_instance_script}
\usage{
ec2_instance_script(
  postgres_password = "postgres",
  phone_number = NA,
  null_user = FALSE
)
}
\arguments{
\item{postgres_password}{Password for postgres database}

\item{phone_number}{Number to sent you a ding. Message is pointless and nondescript}

\item{null_user}{Blank file for nothing to load on server.}
}
\value{
A bash script
}
\description{
ec2_instance_script
}
