% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_instance_start.R
\name{ec2_instance_start}
\alias{ec2_instance_start}
\title{ec2_instance_start}
\usage{
ec2_instance_start(instance_id = NA)
}
\arguments{
\item{instance_id}{An aws ec2 id: i.e., 'i-034e6090b1eb879e7'}
}
\description{
ec2_instance_start
}
