% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_instance_stop.R
\name{ec2_instance_stop}
\alias{ec2_instance_stop}
\title{ec2_instance_stop}
\usage{
ec2_instance_stop(ids)
}
\arguments{
\item{ids}{An aws ec2 id: i.e., 'i-034e6090b1eb879e7'}
}
\description{
ec2_instance_stop
}
