% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_instance_terminate.R
\name{ec2_instance_terminate}
\alias{ec2_instance_terminate}
\title{ec2_instance_terminate}
\usage{
ec2_instance_terminate(ids, force = FALSE)
}
\arguments{
\item{ids}{An aws ec2 id: i.e., 'i-034e6090b1eb879e7'}

\item{force}{Will ignore question about terminate}
}
\description{
ec2_instance_terminate
}
