% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data_gen.R
\name{ec2_prebuilt_script}
\alias{ec2_prebuilt_script}
\title{ec2_prebuilt_script}
\usage{
ec2_prebuilt_script(postgres_password = "postgres", phone_number = NA,
  null_user = FALSE)
}
\arguments{
\item{postgres_password}{Password for postgres database}

\item{phone_number}{Number to sent you a ding. Message is pointless and nondescript}

\item{null_user}{Blank file for nothing to load on server.}
}
\description{
ec2_prebuilt_script
}
