% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{exists_project_model}
\alias{exists_project_model}
\title{Title}
\usage{
exists_project_model(buildingsRdf, projectSubject, namespaces)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
