% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-writing.R
\name{format_iso_8601z}
\alias{format_iso_8601z}
\title{Parse datetime to ISO 8601 format}
\usage{
format_iso_8601z(datetimes)
}
\arguments{
\item{datetimes}{<array> of POSIXct in any timezone.}
}
\value{
<array> of string with ISO 8601 format.
}
\description{
This function parses the format of a POSIXct object to ISO 8601 with Zulu as UTC timezone definition. 
For instance: 2020-01-01 03:00:00 Europe/Madrid --> "2020-01-01T02:00:00Z"
}
