% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{gaMonitor2}
\alias{gaMonitor2}
\title{Genetic algorithm core monitor}
\usage{
gaMonitor2(object, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{<ga object> GA optimization object}
}
\description{
Callback function called during the genetic algorithm
execution in order to monitor optimization evolution
}
