% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{generate_cross_sectional_benchmarking_indicator}
\alias{generate_cross_sectional_benchmarking_indicator}
\title{Title}
\usage{
generate_cross_sectional_benchmarking_indicator(
  data,
  isReal,
  indicator,
  frequency,
  groupSubject,
  prefixUtcTimeColumns,
  indicatorsUnitsSubjects,
  prevResults = NULL
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
