% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{generate_eem_assessment_indicators}
\alias{generate_eem_assessment_indicators}
\title{Title}
\usage{
generate_eem_assessment_indicators(
  data,
  indicators,
  indicatorsNotAggregableByTime,
  measuredProperty,
  measuredPropertyComponent,
  frequencies,
  buildingId,
  buildingSubject,
  timeColumn,
  localTimeZone,
  eemProjectDf,
  consumptionColumn,
  indicatorsTimeAggregationFunctions,
  indicatorsUnitsSubjects,
  baselineConsumptionColumn = NULL,
  buildingGrossFloorArea = NA,
  carbonEmissionsColumn = NULL,
  energyPriceColumn = NULL,
  modelName = NULL,
  modelId = NULL,
  modelLocation = NULL,
  modelStorageInfrastructureSubject = NULL,
  modelTypeSubject = NULL,
  estimateWhenAggregate = T,
  prevResults = NULL
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
