% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{generate_longitudinal_benchmarking_indicators}
\alias{generate_longitudinal_benchmarking_indicators}
\title{Title}
\usage{
generate_longitudinal_benchmarking_indicators(
  data,
  indicators,
  measuredProperty,
  measuredPropertyComponent,
  frequencies,
  buildingId,
  buildingSubject,
  timeColumn,
  localTimeZone,
  consumptionColumn,
  indicatorsUnitsSubjects,
  baselineConsumptionColumn = NULL,
  buildingGrossFloorArea = 0,
  outdoorTemperatureColumn = NULL,
  heatingDegreeDays18Column = NULL,
  coolingDegreeDays21Column = NULL,
  carbonEmissionsColumn = NULL,
  energyPriceColumn = NULL,
  modelName = NULL,
  modelId = NULL,
  modelLocation = NULL,
  modelStorageInfrastructureSubject = NULL,
  modelTypeSubject = NULL,
  modelBaselineYear = NULL,
  estimateWhenAggregate = T,
  prevResults = NULL
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
