% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_all_buildings_list}
\alias{get_all_buildings_list}
\title{Get all building subjects}
\usage{
get_all_buildings_list(buildingsRdf)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{buildingSubjects}{<array> of URIs of the building subjects.}
}
\value{
<array> with the subject URI for each building.
}
\description{
This function get all building subjects available from a BIGG-harmonised dataset.
}
