% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_analytical_groups_by_building_subject}
\alias{get_analytical_groups_by_building_subject}
\title{Title}
\usage{
get_analytical_groups_by_building_subject(buildingsRdf)
}
\arguments{
\item{arg}{<>}
}
\description{
Description
}
