% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_area_building}
\alias{get_area_building}
\title{Get gross floor area of a building}
\usage{
get_area_building(buildingsRdf, buildingSubjects)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{buildingSubjects}{<array> of URIs of the building subjects.}
}
\value{
named <array> with the total gross floor area for each building.
}
\description{
This function get from a BIGG-harmonised dataset the gross floor area of a list of buildings.
}
