% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_building_eems}
\alias{get_building_eems}
\title{Get the Energy Efficiency Measures (EEMs) subjects from a set of buildings}
\usage{
get_building_eems(buildingsRdf, buildingSubjects = NULL)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{buildingSubjects}{<array> of URIs related with the building subjects.}
}
\value{
<data.frame> containing the building subject, building element and 
EEM subject.
}
\description{
This function search for all the available EEMs subjects
in a set of buildings. It also relates the EEMs with the 
building element.
}
