% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_building_identifiers}
\alias{get_building_identifiers}
\title{Calculate the building identifiers}
\usage{
get_building_identifiers(buildingSubjects)
}
\arguments{
\item{buildingSubjects}{<array> of URIs of the building subjects.}
}
\description{
This function calculates the building identifiers from a list of building subjects.
}
