% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{get_building_namespaces}
\alias{get_building_namespaces}
\title{Title}
\usage{
get_building_namespaces(buildingsRdf, buildingSubjects)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
