% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{get_buildings_subjects}
\alias{get_buildings_subjects}
\title{Title}
\usage{
get_buildings_subjects(buildingsRdf, buildingIdsFromOrganization)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
