% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{get_change_point_temperature}
\alias{get_change_point_temperature}
\title{Estimate the optimal change point temperature and weather dependance}
\usage{
get_change_point_temperature(
  consumptionData,
  weatherData,
  consumptionFeature,
  temperatureFeature,
  consumptionGroupFeature = NULL,
  localTimeZone,
  plot = F
)
}
\arguments{
\item{consumptionData}{<data.frame> containing consumption time series.
The time zone of the datetimes must be UTC.}

\item{weatherData}{<data.frame> containing temperature time series.
The time zone of the datetimes must be UTC.}

\item{consumptionFeature}{<string> with the name of consumption column.}

\item{temperatureFeature}{<string> with the name of temperature column.}

\item{consumptionGroupFeature}{<string> with the name of column in consumptionData
that describes the grouping feature used. By default, no group is}

\item{localTimeZone}{<string> specifying the local time zone related to
the building in analysis. The format of this time zones are defined by
the IANA Time Zone Database (https://www.iana.org/time-zones). This
argument is optional, by default no transformation to local time zone is
done.}

\item{plot}{<boolean> Plot change point model describing load vs
temperature relationship in order to identify breakpoints}
}
\value{
Changepoint analysis results attributes <struct>
    - group. Group unique identifier
    - tbalh. Change point temperature in heating season
    - tbalh. Change point temperature in cooling season
    - heating. Heating dependency detected
    - cooling. Cooling dependency detected
}
\description{
Calculate change point temperature by modeling temperature vs consumption
relationship using a multi-step weather dependency signature model.
Multi-step weather dependency signature model is based in best
fitting penalised regression model analysis. Regression quality analysis
is used to identify wether proposed model properly describes the
weather vs consumption relationship. Model coefficients analysis is used
is used to identify type and amount of weather dependency.
}
