% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_device_aggregators}
\alias{get_device_aggregators}
\title{Get the metadata and time series from a filtered set of device aggregators}
\usage{
get_device_aggregators(
  buildingsRdf,
  timeseriesObject = NULL,
  allowedBuildingSubjects = NULL,
  allowedMeasuredProperties = NULL,
  useEstimatedValues = F,
  ratioCorrection = T,
  containsEEMs = F,
  alignGreaterThanHourlyFrequencyToYearly = F,
  transformToAggregatableMeasuredProperty = F,
  aggregatableMeasuredPropertyName = NULL,
  measuredPropertiesToAggregate = NULL,
  defaultFactorsByMeasuredProperty = NULL
)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{timeseriesObject}{<string> path of JSON files, or <list> of time series.}

\item{allowedBuildingSubjects}{<array> of URI(s) containing the allowed building 
subject(s).}

\item{allowedMeasuredProperties}{<array> of string(s) containing the allowed 
measured property(ies).}

\item{useEstimatedValues}{<boolean> describing if the estimated values of time 
series should be taken into account.}

\item{ratioCorrection}{<boolean> describing whether a ratio correction should 
be done, or not. Important to set to TRUE when time series contain gaps.}

\item{containsEEMs}{<boolean> to filter only those buildings that contain one
or more EEM.}

\item{alignGreaterThanHourlyFrequencyToYearly}{<boolean> to force time
alignment to P1Y frequency if original alignment frequency is greater to PT1H.}
}
\value{
<list> of time series and metadata related with all device aggregators.
}
\description{
This function get the metadata and time series from a filtered set of device 
aggregators available in a BIGG-harmonised dataset.
}
