% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{get_device_aggregators_by_building}
\alias{get_device_aggregators_by_building}
\title{Title}
\usage{
get_device_aggregators_by_building(
  buildingsRdf,
  timeseriesObject = NULL,
  allowedBuildingSubjects = NULL,
  allowedMeasuredProperties = NULL,
  useEstimatedValues = F,
  ratioCorrection = T,
  containsEEMs = F,
  alignGreaterThanHourlyFrequencyToYearly = F
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
