% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_device_aggregators_metadata}
\alias{get_device_aggregators_metadata}
\title{Get the metadata of all device aggregators
 
This function gets all the device aggregators available in a BIGG-harmonised dataset.
It also provides metadata with the characteristics of this device aggregators.}
\usage{
get_device_aggregators_metadata(buildingsRdf)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}
}
\description{
Get the metadata of all device aggregators
 
This function gets all the device aggregators available in a BIGG-harmonised dataset.
It also provides metadata with the characteristics of this device aggregators.
}
