% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_eem_details}
\alias{get_eem_details}
\title{Get the complete metadata from a set of Energy Efficiecy Measures (EEMs)}
\usage{
get_eem_details(buildingsRdf, eemSubjects = NULL)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{eemSubjects}{<array> of URIs related with the EEM subjects.}
}
\value{
<data.frame> containing the metadata of EEMs
}
\description{
This function get the metadata of a set of EEMs in a BIGG-harmonised dataset. 
The investment cost is always converted to Euros, for benchmarking purposes.
}
