% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-writing.R
\name{get_eem_lifespan}
\alias{get_eem_lifespan}
\title{Get the Energy Efficiency Measures (EEM) lifespan}
\usage{
get_eem_lifespan(lifespans = NULL, eemTypes)
}
\arguments{
\item{lifespans}{a named <array> defining the default lifespans to consider for each type of EEM.}

\item{eemTypes}{an <array> that defines the EEM types.}
}
\value{
<array> of lifespans for each EEM type.
}
\description{
From a list of lifespans, obtain the list of lifespans of a set of EEM types
}
