% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-writing.R
\name{get_eem_measured_property_components}
\alias{get_eem_measured_property_components}
\title{Get the Energy Efficiency Measures (EEM) measured property components}
\usage{
get_eem_measured_property_components(
  componentsPerEem = NULL,
  eemTypes,
  prefixColumns = ""
)
}
\arguments{
\item{componentsPerEem}{a named <list> defining the default measured property components
to consider for each type of EEM.}

\item{eemTypes}{an <array> that defines the EEM types.}
}
\value{
<data.frame> of measured components for each EEM type.
}
\description{
From a list of measured property components, obtain the logical matrix of measured property components
for each EEM.
}
