% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_emissions_metadata}
\alias{get_emissions_metadata}
\title{Get the emissions metadata}
\usage{
get_emissions_metadata(buildingsRdf, sensorId)
}
\arguments{
\item{buildingsRdf}{<rdf> containing all metadata about buildings. 
It must be harmonised to BIGG Ontology.}

\item{sensorId}{<string> with the sensor identifier 
(e.g. related to some consumption time series).}
}
\value{
<data.frame> with the emissions metadata.
}
\description{
This function extract the emissions metadata related to a certain sensor identifier.
}
