% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_geonames_attr_from_buildings}
\alias{get_geonames_attr_from_buildings}
\title{Get GEONAMES attributes from buildings}
\usage{
get_geonames_attr_from_buildings(buildingsRdf, buildingSubjects)
}
\description{
Get GEONAMES attributes from buildings
}
