% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_KPI_by_building}
\alias{get_KPI_by_building}
\title{Title}
\usage{
get_KPI_by_building(
  buildingsRdf,
  timeseriesObject,
  buildingSubject,
  KPI,
  frequency,
  ratedBy,
  localTz
)
}
\arguments{
\item{arg}{<>}
}
\description{
Description
}
