% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{get_KPI_timeseries}
\alias{get_KPI_timeseries}
\title{Get KPIs time series from harmonised data}
\usage{
get_KPI_timeseries(
  buildingsRdf,
  timeseriesObject,
  buildingSubject,
  name,
  fromModel,
  frequency
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
