% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{get_local_min_from_density}
\alias{get_local_min_from_density}
\title{Title}
\usage{
get_local_min_from_density(
  values,
  onlyPosition = NULL,
  minNumberValleys = 1,
  kDensityAdjuster = 3
)
}
\arguments{
\item{arg}{<>}
}
\description{
Description
}
