% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_matrix_building_space_use_types}
\alias{get_matrix_building_space_use_types}
\title{Title}
\usage{
get_matrix_building_space_use_types(buildingsRdf, splitSublevels = T)
}
\arguments{
\item{arg}{<>}
}
\description{
Description
}
