% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_measurements_by_building_metadata}
\alias{get_measurements_by_building_metadata}
\title{Get the metadata of all measurements by building
 
This function gets all the measurements available in a BIGG-harmonised dataset.
It also provides metadata with the time series characteristics of these measurements 
and their relation with building, building space, data provider and sensor concepts.}
\usage{
get_measurements_by_building_metadata(buildingsRdf)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}
}
\description{
Get the metadata of all measurements by building
 
This function gets all the measurements available in a BIGG-harmonised dataset.
It also provides metadata with the time series characteristics of these measurements 
and their relation with building, building space, data provider and sensor concepts.
}
