% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_sensor_file}
\alias{get_sensor_file}
\title{Read time series files and generate a list of them split by sensor identifiers}
\usage{
get_sensor_file(timeseriesObject, sensorId)
}
\arguments{
\item{timeseriesObject}{<string> describing the local path with the JSON 
files or a <list> containing the time series.}

\item{sensorId}{<string> identifying a time series.}
}
\value{
<list> of data.frames with all time series found.
}
\description{
This function gets time series from JSON files that has the name of a 
selected sensor identifier. If a list of time series is provided, this function
only filters the input list considering the sensor identifiers.
}
