% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_sensor_metadata}
\alias{get_sensor_metadata}
\title{Get sensor metadata}
\usage{
get_sensor_metadata(buildingsRdf, sensorId, tz)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{sensorId}{<uri> or directly the hash of a measurement.}

\item{tz}{<string> specifying the local time zone related to the
building in analysis. The format of this time zones are defined by the IANA
Time Zone Database (https://www.iana.org/time-zones).}
}
\value{
<data.frame> with the metadata of the sensor identifier.
}
\description{
This function gets the available metadata of a certain sensor time series.
}
