% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_std_regioncode_from_buildings}
\alias{get_std_regioncode_from_buildings}
\title{Get ISO 3166-2 Region Codes from buildings using Country Code and Province information}
\usage{
get_std_regioncode_from_buildings(countryCode, Province)
}
\description{
Get ISO 3166-2 Region Codes from buildings using Country Code and Province information
}
