% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_tz_building}
\alias{get_tz_building}
\title{Get timezone of a building}
\usage{
get_tz_building(buildingsRdf, buildingSubjects)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{buildingSubjects}{<array> of URIs of the building subjects.}
}
\value{
named <array> with the timezone for each building. 
The format of this time zones are defined by
the IANA Time Zone Database (https://www.iana.org/time-zones).
}
\description{
This function get from a BIGG-harmonised dataset the timezone of a list of buildings.
}
