% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{getAttribute}
\alias{getAttribute}
\title{Get single attribute from each of the features in the features list}
\usage{
getAttribute(features, name)
}
\arguments{
\item{features}{<list> List of features to get attributes from
Feature description
feature = list(
   datatype = <type of data> (discrete | integer | float)
   min = <min value>,
   max = <max value>,
   nlevels = <number of levels>,
   levels = <levels> (optional: discrete type specific)
)}

\item{name}{<string> Name of the attribute to get}
}
\value{
<list> of features with the single attribute requested
}
\description{
Get single attribute from each of the features in the features list
}
