\name{Glycolysis}
\docType{data}
\alias{Glycolysis}
\title{ Metabolic reconstruction of Glycolysis pathway }
\description{Model of Glycolysis pathway}
\usage{Glycolysis}
\format{A model file created as an example}
\references{ Gavai et al, 
\emph{BiGGR: An open source initiative for constraint-based modeling in R using the BiGG database}. }
\keyword{saureus}