\name{gnel}
\encoding{latin1}
\docType{data}
\alias{gnel}
\title{ Example graphnel object of Glycolysis pathway }
\description{"Glycolysis"}
\usage{gnel}
\format{A graphnel object file created as an example}
\references{ Gavai et al, "BiGGR: An open source initiative for constraint-based modeling in R using the BiGG database". }
\keyword{Glycolysis}