\name{H.pylori_ilT341}
\docType{data}
\alias{H.pylori_ilT341}
\title{H.pylori in silico genome-scale characterization of single and double deletion mutants }
\description{
An Expanded Metabolic Reconstruction of Helicobacter pylori (iIT341 GSM/GPR): An in silico genome-scale characterization of single and double deletion mutants
}
\usage{H.pylori_ilT341}
\format{An sbml object of class rsbml}
\source{Journal of Bacteriology, 187(16): 5818-5830 (2005)}
\references{
Thiele, I., Vo, T.D., Price, N.D. and Palsson, B., \emph{An Expanded Metabolic Reconstruction of Helicobacter pylori} (iIT341 GSM/GPR): An in silico genome-scale characterization of single and double deletion mutants", Journal of Bacteriology, 187(16): 5818-5830 (2005)
}

\keyword{datasets}
