\name{H.sapiens_Recon_1}
\docType{data}
\alias{H.sapiens_Recon_1}
\title{ Human metabolic reconstruction }
\description{ "Global reconstruction of the human metabolic network based on genomic and bibliomic data"}
\usage{H.sapiens_Recon_1}
\format{An sbml object of class rsbml}
\source{Proc. Natl Acad. Sci. 104(6):1777-82 (2007)}
\references{
Duarte, N.D., Becker, S. A., Jamshidi, N., Thiele, I., Mo, M. L., Vo, T. D., Srivas, R., Palsson, B. O., "Global reconstruction of the human metabolic network based on genomic and bibliomic data", Proc. Nat Acad. Sci. 104(6):1777-82 (2007)
}
\keyword{hsapiens}
