% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{hourly_timesteps}
\alias{hourly_timesteps}
\title{Time casting from hours to ISO 8601 timesteps}
\usage{
hourly_timesteps(nHours, original_timestep)
}
\arguments{
\item{nHours}{<integer> Number of hours to obtain}

\item{original_timestep}{<string> in ISO 8601 format representing the period
or timestep (e.g. "PT15M","PT1H", "P3M", "P1D" ,...).}
}
\value{
<integer> corresponding to the number of hours
}
\description{
Supported ISO 8601 timesteps:
  - PT1S
  - PT1M
  - PT1H
  - P1D
  - P1W
  - P1M
  - P1Y
}
