% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{hyperparameters_tuning}
\alias{hyperparameters_tuning}
\title{Hyperparameters tuning}
\usage{
hyperparameters_tuning(
  opt_criteria,
  opt_function,
  features,
  suggestions = NULL,
  keepBest = TRUE,
  parallel = FALSE,
  monitor = TRUE,
  ...
)
}
\description{
Hyperparameter tuning via model fitting optimization. See optimize
documentation for more details
}
