% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python.R
\name{install_python}
\alias{install_python}
\title{install_python: sets up python environment for awsR}
\usage{
install_python(method = "auto", conda = "auto", envname = "r_reticulate")
}
\arguments{
\item{method}{method argument for py_install}

\item{conda}{conda argument for py_install}

\item{envname}{a conda or python virtual environment name}
}
\description{
install_python: sets up python environment for awsR
}
