% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_utilities.R
\name{instance_ec2_create}
\alias{instance_ec2_create}
\title{instance_ec2_create}
\usage{
instance_ec2_create(resource, ImageId = NA, InstanceType = "t2.nano",
  min = 1, max = 1)
}
\arguments{
\item{resource}{The result of resource_ec2}

\item{ImageId}{An aws ec2 id: i.e., 'ami-0174e69c12bae5410'}

\item{InstanceType}{See \url{https://aws.amazon.com/ec2/instance-types/}}
}
\description{
instance_ec2_create
}
