% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_utilities.R
\name{instance_ec2_stop}
\alias{instance_ec2_stop}
\title{instance_ec2_create}
\usage{
instance_ec2_stop(resource, ids, terminate = FALSE)
}
\arguments{
\item{resource}{The result of resource_ec2}

\item{ids}{An aws ec2 id: i.e., 'i-034e6090b1eb879e7'}

\item{terminate}{An boolean to specify whether to stop or terminate}
}
\description{
instance_ec2_create
}
