% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{iso8601_period_to_text}
\alias{iso8601_period_to_text}
\title{ISO 8601 period to natural text}
\usage{
iso8601_period_to_text(x, only_first = F)
}
\arguments{
\item{x}{<string> defining the frequency to be converted. 
It must follow ISO 8601 format representing the time step.}

\item{only_first}{<boolean> specifying if only the first 
timestep item should be converted.}
}
\value{
<string> with the time step in natural text.
}
\description{
This function converts an ISO 8601 period (e.g. PT1H30M) to natural text (in the example, "1 hour, 30 mins").
}
