% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{iso8601_period_to_timedelta}
\alias{iso8601_period_to_timedelta}
\title{ISO 8601 period to period}
\usage{
iso8601_period_to_timedelta(x)
}
\arguments{
\item{x}{<string> defining the frequency to be converted. 
It must follow ISO 8601 format representing the time step.}
}
\value{
<period> according the format defined in lubridate's 'Period-class'.
}
\description{
This function converts an ISO 8601 period (e.g. PT1H30M) to a period object.
}
