% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyfile_create.R
\name{keyfile_create}
\alias{keyfile_create}
\title{keyfile_create}
\usage{
keyfile_create(keyname = NA, save_to_directory = TRUE)
}
\arguments{
\item{keyname}{name for keyfile}

\item{save_to_directory}{save to working directory}
}
\description{
keyfile_create
}
