% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{lag_components}
\alias{lag_components}
\title{This function shift in time a set of features in order to be used in
the training and prediction of the models. It is an important step
for the multi-step prediction of Autoregressive models, where the estimated
output is directly used in the subsequent predictions.}
\usage{
lag_components(
  data,
  maxLag,
  featuresNames = NULL,
  predictionStep = NULL,
  forceGlobalInputFeatures = NULL,
  forceInitInputFeatures = NULL,
  forceInitOutputFeatures = NULL,
  fillInitNAs = F
)
}
\arguments{
\item{data}{<timeSeries> Containing the multiple series to transform. Optionally,
other variables that are not declared in featuresNames can be bypassed
to the output}

\item{maxLag}{<integer> Describing the maximum lags to be considered. One
feature will be generated for each lag}

\item{featuresNames}{<list> selecting the series to transform.}

\item{forceGlobalInputFeatures}{<dict>}

\item{forceInitInputFeatures}{<dict>}

\item{forceInitOutputFeatures}{<dict>}

\item{fillInitNAs}{<boolean> indicating if the unknown lags should be filled with their last known value.}

\item{predictStep}{<integer> informing of the timestep considered in the prediction. 
Only used in prediction mode, when training it doesn't need to be described}
}
\value{
data <timeSeries> containing the same initial information of the data
input argument, plus the lagged components as new columns.
}
\description{
This function shift in time a set of features in order to be used in
the training and prediction of the models. It is an important step
for the multi-step prediction of Autoregressive models, where the estimated
output is directly used in the subsequent predictions.
}
