% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{lpf_ts}
\alias{lpf_ts}
\title{This function computes the first-order low pass filter for smoothing
a time series}
\usage{
lpf_ts(
  data,
  featuresNames,
  smoothingTimeScaleParameter,
  outputFeaturesNames = NULL,
  inplace = T,
  autoUnbox = F
)
}
\arguments{
\item{data}{<timeSeries> containing the series to transform. Optionally,
other variables that are not declared in featuresNames can be bypassed to
the output}

\item{featuresNames}{<string> selecting the series to transform}

\item{smoothingTimeScaleParameter}{<float> of the smoothing time scale
parameter. It corresponds to the so-called alpha parameter of low pass
filter. It ranges from 0 to 1.}

\item{inplace:}{<boolean> indicating if the output should be the original data argument, 
plus the transformed objects -True- , or only the transformed series -False.}

\item{autoUnbox:}{<boolean> indicating if the output timeSeries should be unboxed of the 
resultant data.frame, obtaining a numeric vector. Only is usable when inplace is True.}
}
\value{
data <timeSeries> containing the same initial information of the
data input argument, but updating the featuresNames variables to the
low-pass filtered version
}
\description{
This function computes the first-order low pass filter for smoothing
a time series
}
