\name{M.barkeri_iAF692}
\docType{data}
\alias{M.barkeri_iAF692}
\title{ Metabolic reconstruction of M.barkeri }
\description{"Modeling methanogenesis with a genome-scale metabolic reconstruction of Methanosarcina barkeri"}
\usage{M.barkeri_iAF692}
\format{An sbml object of class rsbml}
\source{Molecular Systems Biology, 2(1):msb4100046-E1-E14 (2006)}
\references{
Feist, A.M., Scholten, J.C.M., Palsson, B.O., Brockman, F.J., and Ideker, T., "Modeling methanogenesis with a genome-scale metabolic reconstruction of Methanosarcina barkeri", Molecular Systems Biology, 2(1):msb4100046-E1-E14 (2006)
}
\keyword{mbarkeri}
