\name{M.tuberculosis_iNJ661}
\docType{data}
\alias{M.tuberculosis_iNJ661}
\title{Metabolic reconstruction of M.tuberculosis from the BiGG database}
\description{
  A metabolic reconstruction for tuberculosis.
  The dataset was generated by downloading the SBML file of the
reconstruction (http://bigg.ucsd.edu/bigg/exportSelect.pl) which was subsequently converted
into an object of class \code{SBML} using the \code{rsbml_read} function
from the \code{rsbml} package.
}
\usage{data(M.tuberculosis_iNJ661)}
\format{An sbml object of class rsbml}

\details{Note that the files in the BiGG database fail the unit
  consistancy check of the \code{rsbml_read} function. To avoid
  unit checking when creating \code{SBML} objects, the substance units in the reaction tags
  were parsed out from the database SBML files (see example below).}

\source{http://bigg.ucsd.edu/bigg/exportSelect.pl}
\references{
Feist, A.M., Henry, C.S., Reed, J.L., Krummenacker, M., Joyce, A.R., Karp, P.D., Broadbelt, L.J., Hatzimanikatis, V., Palsson, B.O., 
\emph{A genome-scale metabolic reconstruction for Escherichia coli K-12 MG1655 that accounts for 1260 ORFs and thermodynamic information}, olecular Systems Biology, 3:121 (2007)
}


\examples{

\dontrun{
##The dataset was generated as follows:
##SBML_export.xml was downloaded from http://bigg.ucsd.edu/bigg/exportSelect.pl
##and a newline was added at the end of the file
file <- "SBML_export.xml"
string <- paste(readLines(file), collapse="\n")
##Parse out units to avoid validation error
string <- gsub("units=\".+?\"", "", string)
M.tuberculosis_iNJ661 <- rsbml_read(text=string) 
}

##load data and get all reaction IDs
data(M.tuberculosis_iNJ661)
model <- M.tuberculosis_iNJ661@model
##get all reaction identifiers
sapply(model@reactions, id)


}

\keyword{tuberculosis}
