% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{make.affinity}
\alias{make.affinity}
\title{Calculate affinity}
\usage{
make.affinity(S, n.neighboors = 2)
}
\arguments{
\item{S}{<matrix> Matrix to evaluate}

\item{n.neighboors}{<int> Number of neighboors}
}
\value{
Affinity matrix
}
\description{
Code from https://www.di.fc.ul.pt/~jpn/r/spectralclustering/spectralclustering.html
Done applying a k-nearest neighboor filter to build a representation of a
graph. connecting just the closest dataset points. However, to be symmetric,
if Aij is selected as a nearest neighboor, so will Aji.
}
