% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{make.similarity}
\alias{make.similarity}
\title{Calculate similarity matrix}
\usage{
make.similarity(my.data, similarity)
}
\arguments{
\item{my.data}{<matrix> Matrix to evaluate}

\item{similariry}{<function> Similarity function}
}
\value{
Similarity matrix
}
\description{
Code from https://www.di.fc.ul.pt/~jpn/r/spectralclustering/spectralclustering.html
Similarity Matrix, S, as the matrix that at Sij=s(xi,xj) gives the
similarity between observations xi and xj.
Common similarity measures are euclidean (default) an gaussian
}
