% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-writing.R
\name{namespace_integrator}
\alias{namespace_integrator}
\title{Namespace integrator}
\usage{
namespace_integrator(items, namespaces = NULL)
}
\arguments{
\item{items}{<array> of strings with the URIs to complete.}

\item{namespaces}{named <array>. Example: 'biggr::bigg_namespaces'.}
}
\value{
<array> of strings with the complete namespace integrated.
}
\description{
This function converts the simplified namespace subtring of multiple URIs to his complete URI namespace.
}
