% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{normalise_dlc}
\alias{normalise_dlc}
\title{normalise load}
\usage{
normalise_dlc(
  data,
  localTimeZone,
  transformation = "relative",
  inputVars = c("load_curves"),
  nDayParts = 24,
  holidays = NULL
)
}
\arguments{
\item{localTimeZone}{<string> timezone}

\item{transformation}{<string>  data transformation required.
Both "relative" and "absolute" supported. By default "absolute"}

\item{inputVars}{<list of strings> Select a number of features as
output for further steps}

\item{nDayParts}{<int> number of part days}

\item{holidays}{<list date> holidays dates}

\item{<timeserie>}{data}
}
\value{
normalised load
}
\description{
normalise load
}
