% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{paramiko}
\alias{paramiko}
\title{paramiko module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
paramiko
}
\description{
paramiko module
}
\keyword{datasets}
