% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{parse_device_aggregator_formula}
\alias{parse_device_aggregator_formula}
\title{Title}
\usage{
parse_device_aggregator_formula(
  buildingsRdf,
  timeseriesObject,
  buildingSubject,
  formula,
  inputFrequency,
  outputFrequency,
  aggFunctions,
  useEstimatedValues,
  ratioCorrection = F
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
