% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_instance_create.R
\name{rds_instance_create}
\alias{rds_instance_create}
\title{rds_instance_create}
\usage{
rds_instance_create(
  db_instance_identifier = NA,
  allocated_storage = 10L,
  database_name = NA,
  engine = "postgres",
  storage_type = "gp2",
  auto_minor_version_upgrade = TRUE,
  multi_az = FALSE,
  master_username = NA,
  master_user_password = NA,
  database_instance_class = "db.t2.micro"
)
}
\arguments{
\item{db_instance_identifier}{NA,}

\item{allocated_storage}{10L,}

\item{database_name}{NA,}

\item{engine}{"postgres",}

\item{storage_type}{"gp2",}

\item{auto_minor_version_upgrade}{TRUE,}

\item{multi_az}{FALSE,}

\item{master_username}{NA,}

\item{master_user_password}{NA,}

\item{database_instance_class}{"db.t2.micro"}
}
\description{
rds_instance_create
}
