% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{read_and_transform_sensor}
\alias{read_and_transform_sensor}
\title{Title}
\usage{
read_and_transform_sensor(
  timeseriesObject,
  buildingsRdf,
  sensorId,
  tz,
  outputFrequency,
  aggFunctions,
  useEstimatedValues
)
}
\arguments{
\item{arg}{<>}
}
\value{

}
\description{
Description
}
