\name{Recon2}
\docType{data}
\alias{Recon2}
\title{ Human metabolic reconstruction }
\description{ "A community-driven global reconstruction of human metabolism"}
\usage{Recon2.rda}
\format{An sbml object of class rsbml}
\source{Nature Biotechnology (2013)}
\references{
Thiele I, Swainston N, et al.,"A community-driven global reconstruction of human metabolism", Nature Biotechnology, advanced online publication 3 March 2013
}
\keyword{hsapiens}
