% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{relative_out_of_boundaries_in_periodogram}
\alias{relative_out_of_boundaries_in_periodogram}
\title{Check the ratio of the periodogram out of the boundaries}
\usage{
relative_out_of_boundaries_in_periodogram(ts, taper = 0.1)
}
\arguments{
\item{ts}{<vector> A univariate time series of values, usually the training residuals of a model}

\item{taper}{<numeric> A proportion tapered in forming the periodogram.}
}
\value{
The proportion of the periodogram out of boundaries.
}
\description{
Check the ratio of the periodogram out of the boundaries
}
