% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{resample}
\alias{resample}
\title{Resample time serie to new ISO 8601 timestep}
\usage{
resample(data, timestep)
}
\arguments{
\item{data}{<timeSeries> Input time series to be resampled}

\item{timeStep}{<string> A string in ISO 8601 format representing the period
or timestep (e.g. "PT15M","PT1H", "P3M", "P1D" ,...).}
}
\value{
Resampled time series
}
\description{
Time serie is upsampled creating synthetic forward fill time serie
in case of higher frequency timestep required
Time serie is downsampled maintaing curent values but
repeated indexes in case of lower frequency timestep required
}
