% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{resample_and_aggregate}
\alias{resample_and_aggregate}
\title{Resample time series to new ISO 8601 timestep}
\usage{
resample_and_aggregate(
  data,
  timestep,
  func,
  minRatioToTimeAggregateIndicators = 0,
  estimateWhenAggregate = T
)
}
\arguments{
\item{data}{<data.frame> describing the input time series to be resampled. 
Time column: 'time', value column: 'value'. If you use weighted average aggregation,
you must include a column in data named 'weights'}

\item{func}{<string> A string defining the function to use in aggregate
("SUM","WEIGHTED-AVG", "AVG").}

\item{minRatioToTimeAggregateIndicators}{<float> Minimum percentage of known values
in a timestep. Default is 0%}

\item{estimateWhenAggregate}{<boolean> defining if the output value should be an
estimator (theorical approx considering the timestep) or the real value (unconsidering gaps).}

\item{timeStep}{<string> A string in ISO 8601 format representing the period
or timestep (e.g. "PT15M","PT1H", "P3M", "P1D" ,...).}
}
\value{
<data.frame> corresponding to the resampled and aggregated time series
}
\description{
Time series is upsampled creating synthetic forward fill time series
in case of higher frequency timestep required. On contrary, is 
downsampled maintaing current values but repeated indexes in case 
of lower frequency timestep required.
}
