% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{round_up}
\alias{round_up}
\title{Round up value or array}
\usage{
round_up(x)
}
\arguments{
\item{x}{<value|array> Value to be rounded}
}
\value{
Rounded value
}
\description{
Round up value or array
}
