\name{S.aureus_iSB619}
\docType{data}
\alias{S.aureus_iSB619}
\title{ Metabolic reconstruction of S.aureus }
\description{"Genome-scale reconstruction of the metabolic network in Staphylococcus aureus N315: an initial draft to the two-dimensional annotation"}
\usage{S.aureus_iSB619}
\format{An sbml object of class rsbml}
\source{BMC Microbiology, 5(1):8 (2005)}
\references{
Becker, S.A. and Palsson, B.O., "Genome-scale reconstruction of the metabolic network in Staphylococcus aureus N315: an initial draft to the two-dimensional annotation", BMC Microbiology, 5(1):8 (2005)
}
\keyword{saureus}
