\name{S.cerevisiae_iND750}
\docType{data}
\alias{S.cerevisiae_iND750}
\title{ Metabolic reconstruction of S.cerevisiae }
\description{"Reconstruction and Validation of Saccharomyces cerevisiae iND750, a Fully Compartmentalized Genome-scale Metabolic Model"}
\usage{S.cerevisiae_iND750}
\format{An sbml object of class rsbml}
\source{Genome Research, 14: 1298-1309 (2004)}
\references{
Duarte, N.C., Herrgard, M.J., and Palsson, B.O., " Reconstruction and Validation of Saccharomyces cerevisiae iND750, a Fully Compartmentalized Genome-scale Metabolic Model", Genome Research, 14: 1298-1309 (2004)
}
\keyword{datasets}
