% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_create_bucket.R
\name{s3_create_bucket}
\alias{s3_create_bucket}
\title{s3_create_bucket}
\usage{
s3_create_bucket(bucket_name = NA, location = "us-east-2")
}
\arguments{
\item{bucket_name}{A name for the bucket}

\item{location}{An AWS region. Defaults to us-east-2}
}
\description{
s3_create_bucket
}
