% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_delete_bucket.R
\name{s3_delete_bucket}
\alias{s3_delete_bucket}
\title{s3_create_bucket}
\usage{
s3_delete_bucket(bucket_name = NA)
}
\arguments{
\item{bucket_name}{A name for the bucket}
}
\description{
s3_create_bucket
}
