% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_delete_file.R
\name{s3_delete_file}
\alias{s3_delete_file}
\title{s3_delete_file}
\usage{
s3_delete_file(bucket, file)
}
\arguments{
\item{bucket}{Bucket to upload to}

\item{file}{File to delete}
}
\description{
s3_delete_file
}
