% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utilities.R
\name{s3_put_object_acl}
\alias{s3_put_object_acl}
\title{s3_put_object_acl}
\usage{
s3_put_object_acl(bucket = NA, file = NA, ACL = "public-read")
}
\arguments{
\item{bucket_name}{bucket_name}
}
\description{
s3_put_object_acl
}
