% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utilities.R
\name{s3_upload_file}
\alias{s3_upload_file}
\title{upload_file}
\usage{
s3_upload_file(bucket, from, to)
}
\arguments{
\item{bucket}{Bucket to upload to}

\item{from}{File to upload}

\item{to}{S3 object name.}
}
\description{
upload_file
}
