% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_group_delete.R
\name{security_group_delete}
\alias{security_group_delete}
\title{security_group_delete}
\usage{
security_group_delete(security_group_id)
}
\arguments{
\item{security_group_id}{a security group ID}
}
\description{
security_group_delete
}
