% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_group_create.R
\name{security_group_revoke}
\alias{security_group_revoke}
\title{security_group_create}
\usage{
security_group_revoke(sg_name = NA, ports = NULL, ips = NULL)
}
\arguments{
\item{ports}{Name of security group}

\item{group_name}{Name of security group}
}
\description{
security_group_create
}
