% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{similarity}
\alias{similarity}
\title{Calculate similarity
Exponential Euclidean distance (Frobenius norm) with weighting factor}
\usage{
similarity(x1, x2, alpha = 1)
}
\arguments{
\item{x1}{<vector> First vector}

\item{x2}{<vector> Second vector}

\item{alpha}{<float> Weighting factor}
}
\value{
Distance between vectors
}
\description{
Calculate similarity
Exponential Euclidean distance (Frobenius norm) with weighting factor
}
