% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{simulate_prediction_intervals}
\alias{simulate_prediction_intervals}
\title{Title}
\usage{
simulate_prediction_intervals(
  predictionResults,
  outputFrequency,
  timeColumn = "time",
  funAggregation = "SUM",
  minPredictionValue = NULL,
  nSim = 1000,
  estimate = F
)
}
\arguments{
\item{arg}{<>}
}
\description{
Description
}
