% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_send_message.R
\name{sns_send_message}
\alias{sns_send_message}
\title{sns_send_message}
\usage{
sns_send_message(
  phone_number = NA,
  message = "Hello World!",
  region = "us-east-1",
  message_aws = FALSE
)
}
\arguments{
\item{phone_number}{Recipient phone number}

\item{message}{Recipient message}

\item{region}{Region to send the message from}

\item{message_aws}{Return AWS data or TRUE}
}
\description{
sns_send_message
}
