% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_master.R
\name{spark_master}
\alias{spark_master}
\title{spark_master}
\usage{
spark_master(
  InstanceType = "t2.medium",
  KeyName = NA,
  SecurityGroupId = NA,
  InstanceStorage = 50
)
}
\arguments{
\item{InstanceType}{'t2.medium'}

\item{KeyName}{= NA}

\item{SecurityGroupId}{= NA}

\item{InstanceStorage}{= 50}
}
\description{
spark_master
}
