% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_slave.R
\name{spark_slave}
\alias{spark_slave}
\title{spark_slave}
\usage{
spark_slave(
  InstanceType = "t2.medium",
  KeyName = NA,
  SecurityGroupId = NA,
  InstanceStorage = 50,
  n_instances = 2,
  master_ip = NULL
)
}
\arguments{
\item{InstanceType}{'t2.medium'}

\item{KeyName}{= NA}

\item{SecurityGroupId}{= NA}

\item{InstanceStorage}{= 50}

\item{n_instances}{= 2}

\item{master_ip}{= NULL}
}
\description{
spark_slave
}
