% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{stage_run_command}
\alias{stage_run_command}
\title{send_file <- function(con, local_path, remote_path) {
  ftp_client = con$open_sftp()
  ftp_client$put(local_path, remote_path)
  ftp_client$close()
}}
\usage{
stage_run_command(
  command = "ls -lah",
  stage_name = "DEV",
  key_filename = "fdren.pem"
)
}
\description{
get_file <- function(con, remote_path, local_path) {
  ftp_client = con$open_sftp()
  ftp_client$get(remote_path, local_path)
  ftp_client$close()
}
}
