% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{subprocess}
\alias{subprocess}
\title{subprocess module}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
subprocess
}
\description{
subprocess module
}
\keyword{datasets}
