% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{sys}
\alias{sys}
\title{sys module}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
sys
}
\description{
sys module
}
\keyword{datasets}
