% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{toBin}
\alias{toBin}
\title{Binary encoding of a value (integer representation)}
\usage{
toBin(x)
}
\arguments{
\item{x}{<integer> Value to be binary encoded}
}
\value{
<integer> representation of binary coded value
}
\description{
Binary encoding of a value (integer representation)
}
