% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data_gen.R
\name{user_data_gen}
\alias{user_data_gen}
\title{user_data_gen}
\usage{
user_data_gen(postgres_password = "postgres", phone_number = NA)
}
\arguments{
\item{postgres_password}{Password for postgres database}

\item{phone_number}{Number to sent you a ding. Message is pointless and nondescript}
}
\description{
user_data_gen
}
