% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{weather_dependence_disaggregator}
\alias{weather_dependence_disaggregator}
\title{Title}
\usage{
weather_dependence_disaggregator(
  predictor,
  df,
  forceNoCooling,
  forceNoHeating,
  forceNoCoolingAndHeating = NULL,
  ...
)
}
\arguments{
\item{arg}{<>}
}
\description{
Description
}
