% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation.R
\name{write_rdf}
\alias{write_rdf}
\title{Export an RDF to Turtle format (TTL)}
\usage{
write_rdf(object, file)
}
\arguments{
\item{object}{<rdf> knowledge graph}

\item{file}{<string> containing the path and filename where to store the data in TTL format.}
}
\value{

}
\description{
Export a knowledge graph, represented in RDF, to a file in Turtle format (TTL)
}
