% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-writing.R
\name{add_item_to_rdf}
\alias{add_item_to_rdf}
\title{Add triplet to and RDF object}
\usage{
add_item_to_rdf(
  object,
  subject,
  classes = NULL,
  dataProperties = NULL,
  objectProperties = NULL,
  namespaces = NULL
)
}
\arguments{
\item{object}{<rdf> base object.}

\item{subject}{<uri> containing the subject of the new item.}

\item{classes}{<uri(s)> containing the subject(s) name(s) in the 
ontology related with this item.}

\item{dataProperties}{<list> containing data properties of the triplet. 
It is literal information, no connections to other subjects of the RDF.}

\item{objectProperties}{<list> containing object properties of the triplet. 
It consists on links to other items of the RDF.}

\item{namespaces}{named <array> that relates simple namespaces and complete 
ones.}
}
\value{
<rdf> object with the new item added.
}
\description{
This function adds a triplet to an RDF object. Classes, data properties and 
object properties can be defined to a certain subject.
}
