% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{append_cost_to_sensor}
\alias{append_cost_to_sensor}
\title{Append the cost and price to some energy time series sensor}
\usage{
append_cost_to_sensor(
  buildingsRdf,
  timeseriesObject,
  tariffSubject,
  measuredProperty,
  frequency,
  energyTimeseriesSensor
)
}
\arguments{
\item{buildingsRdf}{<rdf> containing all metadata about buildings. 
It must be harmonised to BIGG Ontology.}

\item{timeseriesObject}{<string> path of JSON files, or <list> of time 
series.}

\item{tariffSubject}{<uri> with the tariff identifier.}

\item{measuredProperty}{<string> with the energy measured property to 
consider for the tariff cost calculation 
(e.g. EnergyConsumptionGridElectricity, EnergyConsumptionGas)}

\item{frequency}{<string> defining the frequency to be considered in the 
energy cost calculation. It must follow ISO 8601 format representing the 
time step.}

\item{energyTimeseriesSensor}{<data.frame> output from 
get_sensor().}
}
\value{
<data.frame>, by-passing the input argument energyTimeseriesSensor and 
appending columns related to energy cost.
}
\description{
This function calculates the cost and append the price to some 
energy time series sensor, based on a tariff definition for that
sensor in a BIGG-harmonised dataset.
}
