% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{classification_dlc}
\alias{classification_dlc}
\title{Daily load curve classification}
\usage{
classification_dlc(
  data,
  consumptionFeature,
  outdoorTemperatureFeature,
  localTimeZone,
  clustering,
  methodNormalDays = "clusteringCentroids",
  holidaysDatesFeature = NULL,
  abnormalDaysFeature = NULL,
  holidaysDates = c(),
  abnormalDays = c(),
  methodAbnormalDays = "clusteringCentroids"
)
}
\arguments{
\item{data}{<data.frame> containing the time series for total energy
consumption of a building, the outdoor temperature, or whatever input is
needed for clustering.}

\item{consumptionFeature}{<string> containing the column name the consumption
feature in the data argument.}

\item{localTimeZone}{<string> specifying the local time zone related to the
building in analysis. The format of this time zones are defined by the IANA
Time Zone Database (https://www.iana.org/time-zones).}

\item{clustering}{<object> clustering_dlc() output}

\item{outdorTemperatureFeature}{<string> containing the column name of the outdoor
temperature feature in the data argument.}

\item{method:}{<string>. Choose one, by default clusteringCentroids:
absoluteLoadCurvesCentroids: Based on the absolute consumption load
  curve
clusteringCentroids: Based on the inputs considered in the clustering
  procedure. Applying the same transformations done during that process.
classificationModel: Based on a classification model of the calendar
  features.}
}
\value{
dailyClassification: <data.frame>
}
\description{
Classify daily load curves based on the outputs of a clustering and a
new set of data
}
