% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{degree_days}
\alias{degree_days}
\title{Degree days with time aggregation}
\usage{
degree_days(
  data,
  temperatureFeature,
  localTimeZone,
  baseTemperature,
  mode = "heating",
  outputFrequency = "P1D",
  hysteresisBaseTemperature = 2,
  outputFeaturesName = "HDD",
  fixedOutputFeaturesName = F
)
}
\arguments{
\item{data}{<data.frame> containing the series with data. The time zone
of the datetimes must be UTC. The other variables describing the series are
directly bypassed to the output.}

\item{temperatureFeature}{<string> giving the column name of the outdoor
temperature feature.}

\item{localTimeZone}{<string> specifying the local time zone related to
the building in analysis. The format of this time zones are defined by
the IANA Time Zone Database (https://www.iana.org/time-zones). This
argument is optional, by default no transformation to local time zone is
done.}

\item{baseTemperature}{<float> describing the Balance Point
Temperature (BPT) used in the calculation. Below BPT in heating mode,
heat would be required by the building. The contrary in the case of
cooling, over BPT in cooling mode}

\item{mode}{<string> describing the calculation mode, which could be
"cooling" or "heating". By default, "heating" is configured.}

\item{outputFrequency}{<string> The frequency used to resample the daily
degree days. It must be a string in ISO 8601 format representing the
time step. Examples: 'P1D' (One day), 'P1Y' (One year), 'P1M' (One month)...}

\item{outputFeaturesName}{<string> giving the column name used as output of
the degreedays result. By default, "HDD" is configured}

\item{fixedOutputFeaturesName}{<boolean> enable fixed column name used as
output of the degreedays results. Otherwise baseTemperature is added 
as suffix in the output column name}
}
\value{
<data.frame> in the outputTimeStep of the heating or
cooling degree days.
}
\description{
Calculate the degree-days with a desired output frequency and considering
cooling or heating mode.
}
