% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_disruptive_period}
\alias{detect_disruptive_period}
\title{The function detects periods of time with anomalous low consumption
Evaluates multiple time periods in order to find the best one
fitting a model which considers the relationship between
cooling/heating and consumption}
\usage{
detect_disruptive_period(
  data,
  consumptionColumn,
  timeColumn,
  temperatureColumn,
  tz,
  minIniDate,
  maxIniDate,
  minEndDate,
  maxEndDate,
  checkFor = "decrement",
  minDecrementPercentualAffectation = 30,
  minIncrementPercentualAffectation = 60
)
}
\arguments{
\item{data}{<timeSeries> Time serie with potential anomalies in
values}

\item{timeColumn}{<string> Time column in data time serie}

\item{temperatureColumn}{<string> Temperature column in data time serie}

\item{tz}{<string> Timezone}

\item{minIniDate}{<date> Minimum date to start looking for anomalies}

\item{maxIniDate}{<date> Maximum date to start looking for anomalies}

\item{minEndDate}{<date> Minimum date to stop looking for anomalies}

\item{maxEndDate}{<date> Maximum date to stop looking for anomalies}

\item{checkFor}{<string> Which effect will the disruptive
event generate? e.g. the SARS-CoV2 pandemia decrease the energy consumption of tertiary buildings,
so in that case we should check for decrement in consumption 
Options: decrement, increment, incrementAndDecrement.}

\item{consumptionValue}{<string> Consumption column in data time serie}
}
\value{
Period of time <start,end> with anomalous low consumption. Period
of time is within the InitDate and EndDate limits
}
\description{
The function detects periods of time with anomalous low consumption
Evaluates multiple time periods in order to find the best one
fitting a model which considers the relationship between
cooling/heating and consumption
}
