% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_profiled_data}
\alias{detect_profiled_data}
\title{Detect profiled days in data}
\usage{
detect_profiled_data(data, timeColumn = "localtime", valueColumn = "Qe")
}
\arguments{
\item{data}{<data.frame> with potential profiled content.}

\item{timeColumn}{<string> identifying the local time of the 
time series.}

\item{valueColumn}{<string> identifying the value of the time
series.}
}
\value{
<array> of dates that have a high probability to be profiled.
}
\description{
This function detects if there are profiled days in a consumption 
time series.
}
