% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_static_min_max_outliers}
\alias{detect_static_min_max_outliers}
\title{Detect min-max outliers in static data.}
\usage{
detect_static_min_max_outliers(
  data,
  min,
  max,
  includeMin = TRUE,
  includeMax = TRUE
)
}
\arguments{
\item{data}{<array> of floats corresponding to the numerical elements to be evaluated}

\item{min}{<float> The minimum value of the range.}

\item{max}{<float> The maximum value of the range.}

\item{includedMin}{<boolean> An optional argument setting if the
minimum value should be included as valid value (T), or not (F).
Default is T.}

\item{includedMax}{<boolean> An optional argument setting if the maximum
value should be included as valid value (T), or not (F).
Default is T.}
}
\value{
<array> of booleans, representing the outliers of input data argument.
}
\description{
Detect which numerical elements are outside the min-max range.
}
