% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_static_reg_exp}
\alias{detect_static_reg_exp}
\title{Detect reg-exp in static data.}
\usage{
detect_static_reg_exp(data, regExpValues, negativeExp = FALSE)
}
\arguments{
\item{data}{<array> of strings, representing the elements to be evaluated.}

\item{regExpValues}{<array> of regular expressions to be
checked.}

\item{negativeExp}{<boolean> An optional argument to evaluate the inverse
result of the regular expressions. Default is false, thus all elements
that are satified by any of the regular expressions are true.}
}
\value{
<array> of booleans, representing the evaluation results made
for each data item.
}
\description{
Detect which string element satisfy the regular expression
}
