% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_time_step}
\alias{detect_time_step}
\title{The function infers, i.e. automatically deduce from the input data, the
minimum time step (frequency) that can be used for the input time series.}
\usage{
detect_time_step(data)
}
\arguments{
\item{data}{<timeSeries> Input time series whose time step has to be detected.}
}
\value{
timeStep <string> A string in ISO 8601 format representing the period
or timestep (e.g. "PT15M","PT1H", "P3M", "P1D" ,...).
}
\description{
The function infers, i.e. automatically deduce from the input data, the
minimum time step (frequency) that can be used for the input time series.
}
