% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_ts_calendar_model_outliers}
\alias{detect_ts_calendar_model_outliers}
\title{Detect outliers in time series based on a calendar model}
\usage{
detect_ts_calendar_model_outliers(
  data,
  localTimeColumn = "localtime",
  valueColumn = outputName,
  calendarFeatures = c("HOL", "H"),
  mode = "upperAndLower",
  upperModelPercentile = 90,
  lowerModelPercentile = 10,
  upperPercentualThreshold = 30,
  lowerPercentualThreshold = 30,
  holidaysCalendar = c(),
  daysThatAreOutliers = c(),
  window = NULL,
  outputPredictors = F,
  logValueColumn = F,
  autoDetectProfiled = T
)
}
\arguments{
\item{data}{<data.frame> An argument containing the time series from which
the outliers need to be detected.}

\item{holidaysCalendar}{list of dates An optional list giving the dates
where local or national holidays related to the location of the data
argument. Default is empty list.}

\item{daysThatAreOutliers}{list of outlier dates}

\item{window}{string. A string in ISO 8601 format representing the
window (e.g. "P2M","P4M","P14D",...). This is an optional argument setting
the width of the window where the model is trained and evaluated}

\item{localTimeColumn:}{string Local time column name}

\item{valueColumn:}{string Value column name}

\item{calendarFeatures:}{list of strings An optional argument set the
calendar features of the model. Default values are: ["HOL*intercept","H"]}

\item{mode:}{string An optional argument setting which outliers need to be
filtered, the ones upper, or the ones lower to the prediction.
Default is "upperAndLower".}

\item{upperModelPercentile:}{float An optional argument defining the
percentile used in the quantile regression model for the upper prediction.
Default is 90.}

\item{lowerModelPercentile:}{float An optional argument defining the
percentile used in the quantile regression model for the lower prediction.
Default is 10.}

\item{upperPercentualThreshold:}{float It sets the dynamic upper threshold
to detect outliers. It is an optional argument to define the percentage
of difference added to the upper model predition itself. Default is 30.}

\item{lowerPercentualThreshold:}{float It sets the dynamic lower threshold
to detect outliers. An optional argument to define the percentage of
difference substracted to the predition of the model. () Default is 30.}

\item{outputPredictor:}{boolean. Include calendar regression model prediction
results as output}

\item{logValueColumn:}{boolean. Transform value column (log-transformation)}

\item{autoDetectProfiled:}{boolean. Detect and ignore profiled days from
timeseries}
}
\value{
timeSeries with time index outliers mask and optional calendar
regression model prediction
}
\description{
Detect elements of the time series out of a confidence threshold based
on linear model of the calendar variables (month, weekday, hour).
}
