% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{detect_ts_zscore_outliers}
\alias{detect_ts_zscore_outliers}
\title{Detect Z-score outliers in time series.}
\usage{
detect_ts_zscore_outliers(
  data,
  zScoreThreshold,
  window = NULL,
  zScoreExtremesSensitive = TRUE,
  na.rm = T
)
}
\arguments{
\item{data}{<data.frame> describing the input time series whose outliers need 
to be detected. Time column: 'time', value column: 'value'.}

\item{zScoreThreshold}{<float> describing the threshold of the
Z-score calculation.}

\item{window}{<string> in ISO 8601 format representing the window
(e.g. "P7D","P1D", "PT168H" ,...). This is an optional argument setting
the width of the rolling window where the Z-normalization calculation
is considered. This argument allows to adapt the outlier filtering depending
the dynamics of the signal itself. Default value is "NULL", thus no rolling
window is considered.}

\item{zScoreExtremesSensitive}{<boolean> defining if the
aggregation function of the Zscore is the mean (true), or median(false).
The first one makes the Z-score sensitive to extreme values, and the
second no. Default is true.}

\item{na.rm}{<boolean> describing if NAN values should be removed 
in mean, median and sd functions}
}
\value{
<data.frame> with boolean values representing whether a item is an 
outlier, or not.
}
\description{
Detect elements of the time series out of a Z-score threshold,
applied on the whole timeseries or a rolling window of predefined width.
}
