\name{E.coli_iJR904}
\encoding{latin1}
\docType{data}
\alias{E.coli_iJR904}
\title{Ecoli genome-scale model}
\description{
 An expanded genome-scale model of Escherichia coli K-12 (iJR904
 GSM/GPR). The dataset was generated by downloading the SBML file of the
reconstruction (http://bigg.ucsd.edu/bigg/exportSelect.pl) which was subsequently converted
into an object of class \code{SBML} using the \code{rsbml_read} function
from the \code{rsbml} package.
}
\usage{data(E.coli_iJR904)}
\format{An sbml object of class rsbml}

\details{Note that the files in the BiGG database fail the unit
  consistancy check of the \code{rsbml_read} function. To avoid
  unit checking when creating \code{SBML} objects, the substance units in the reaction tags
  were parsed out from the database SBML files (see example below).}

\source{http://bigg.ucsd.edu/bigg/exportSelect.pl}
\references{
Reed, J.L., Vo, T.D., Schilling, C.H., and Palsson, B.O., 
\emph{An expanded genome-scale model of Escherichia coli K-12 (iJR904 GSM/GPR)}, Genome Biology, 4(9): R54.1-R54.12 (2003).
}

\examples{

\dontrun{
##The dataset was generated as follows:
##SBML_export.xml was downloaded from http://bigg.ucsd.edu/bigg/exportSelect.pl
##and a newline was added at the end of the file
file <- "SBML_export.xml"
string <- paste(readLines(file), collapse="\n")
##Parse out units to avoid validation error
string <- gsub("units=\".+?\"", "", string)
E.coli_iJR904 <- rsbml_read(text=string) 
}

##load data and get all reaction IDs
data(E.coli_iJR904)
model <- E.coli_iJR904@model
##get all reaction identifiers
sapply(model@reactions, id)


}

\keyword{ecoli}
