% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{exists_analytical_model}
\alias{exists_analytical_model}
\title{Check if exists an analytical model.}
\usage{
exists_analytical_model(buildingsRdf, modelSubject, namespaces)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{modelSubject}{<uri> of an analytical model.}

\item{namespaces}{named <array> that relates simple namespaces and complete.
ones.}
}
\value{
<boolean> if the model exists.
}
\description{
This function checks if certain analytical model subject exists in a BIGG-harmonised dataset.
}
