% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{exists_project_model}
\alias{exists_project_model}
\title{Check if exists a project.}
\usage{
exists_project_model(buildingsRdf, projectSubject, namespaces)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{projectSubject}{<uri> of project subject.}

\item{namespaces}{named <array> that relates simple namespaces and complete 
ones.}
}
\value{
<boolean> if the project exists.
}
\description{
This function checks if a certain project subject exists in a BIGG-harmonised dataset.
}
