% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_building_namespaces}
\alias{get_building_namespaces}
\title{Get namespaces of a building}
\usage{
get_building_namespaces(buildingsRdf, buildingSubjects)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{buildingSubjects}{<array> of URIs of the building subjects.}
}
\value{
named <array> with the namespace for each building.
}
\description{
This function get from a BIGG-harmonised dataset the namespaces of a list of buildings.
}
