% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_buildings_subjects}
\alias{get_buildings_subjects}
\title{Get building subjects from a set of building identifier from organisation}
\usage{
get_buildings_subjects(buildingsRdf, buildingIdsFromOrganization)
}
\arguments{
\item{buildingsRdf}{<rdf> containing the information of a set of buildings.}

\item{buildingIdsFromOrganization}{<array> of strings containing the building identifiers from the organisation.}
}
\value{
<array> with the subject URI for each building.
}
\description{
This function get the building subjects based on a list of building identifier(s) from organisation.
}
