% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonised-reading.R
\name{get_KPI_timeseries}
\alias{get_KPI_timeseries}
\title{Get one Key Performance Indicator (KPI) time series from a certain building.}
\usage{
get_KPI_timeseries(
  buildingsRdf,
  timeseriesObject,
  buildingSubject,
  name,
  fromModel,
  frequency
)
}
\arguments{
\item{buildingsRdf}{<rdf> containing all metadata about buildings. 
It must be harmonised to BIGG Ontology.}

\item{timeseriesObject}{<array> of strings with paths to JSON files containing time series,
or <list> of time series. It must be harmonised to BIGG Ontology.}

\item{buildingSubject}{<uri> of the building subject in buildingsRdf.}

\item{name}{<string> defining the indicator name to be retrieved.}

\item{fromModel}{<boolean> defining if the time series should be real (FALSE), or estimated (TRUE)}

\item{frequency}{<string> defining the frequency to be retrieved. 
It must follow ISO 8601 format representing the time step. 
Examples: 'P1D' (One day), P1Y' (One year), 'P1M' (One month)}
}
\value{
<data.frame> with columns 'time' and 'value'.
}
\description{
This function use the harmonised data to extract a specific KPI time series of a certain 
building.
}
