\name{H.pylori_ilT341}
\docType{data}
\alias{H.pylori_ilT341}
\title{H.pylori in silico genome-scale characterization of single and double deletion mutants }
\description{
An Expanded Metabolic Reconstruction of Helicobacter pylori (iIT341
GSM/GPR): An in silico genome-scale characterization of single and
double deletion mutants.
The dataset was generated by downloading the SBML file of the
reconstruction (http://bigg.ucsd.edu/bigg/exportSelect.pl) which was subsequently converted
into an object of class \code{SBML} using the \code{rsbml_read} function
from the \code{rsbml} package.

}
\usage{data(H.pylori_ilT341)}
\format{An sbml object of class rsbml}

\details{Note that the files in the BiGG database fail the unit
consistancy check of the \code{rsbml_read} function. To avoid
unit checking when creating \code{SBML} objects, the substance units in the reaction tags
were parsed out from the database SBML files (see example below).}


\source{http://bigg.ucsd.edu/bigg/exportSelect.pl}
\references{
Thiele, I., Vo, T.D., Price, N.D. and Palsson, B., \emph{An Expanded Metabolic Reconstruction of Helicobacter pylori} (iIT341 GSM/GPR): An in silico genome-scale characterization of single and double deletion mutants, Journal of Bacteriology, 187(16): 5818-5830 (2005)
}

\examples{
\dontrun{
##The dataset was generated as follows:
##SBML_export.xml was downloaded from http://bigg.ucsd.edu/bigg/exportSelect.pl
##and a newline was added at the end of the file
file <- "SBML_export.xml"
string <- paste(readLines(file), collapse="\n")
##Parse out units to avoid validation error
string <- gsub("units=\".+?\"", "", string)
H.pylori_ilT341 <- rsbml_read(text=string) 
}

##load data and get all reaction IDs
data(H.pylori_ilT341)
model <- H.pylori_ilT341@model
##get all reaction identifiers
sapply(model@reactions, id)

}



\keyword{hpylori}
