\name{H.sapiens_Recon_1}
\docType{data}
\alias{H.sapiens_Recon_1}
\title{ Reconstruction of human metabolism from the BiGG database }
\description{The dataset was generated by downloading the SBML file of the
reconstruction (http://bigg.ucsd.edu/bigg/exportSelect.pl) which was subsequently converted
into an object of class \code{SBML} using the \code{rsbml_read} function
from the \code{rsbml} package.}
\usage{data(H.sapiens_Recon_1)}
\format{An sbml object of class rsbml}

\details{Note that the files in the BiGG database fail the unit
consistancy check of the \code{rsbml_read} function. To avoid
unit checking when creating \code{SBML} objects, the substance units in the reaction tags
were parsed out from the database SBML files (see example below).}

\source{http://bigg.ucsd.edu/bigg/exportSelect.pl}

\references{
Duarte, N.D., Becker, S. A., Jamshidi, N., Thiele, I., Mo, M. L., Vo, T. D., Srivas, R., Palsson, B. O., Global reconstruction of the human metabolic network based on genomic and bibliomic data, Proc. Nat Acad. Sci. 104(6):1777-82 (2007)
}


\examples{

\dontrun{
##The dataset was generated as follows:
##SBML_export.xml was downloaded from http://bigg.ucsd.edu/bigg/exportSelect.pl
##and a newline was added at the end of the file
file <- "SBML_export.xml"
string <- paste(readLines(file), collapse="\n")
##Parse out units to avoid validation error
string <- gsub("units=\".+?\"", "", string)
H.sapiens_Recon_1 <- rsbml_read(text=string) 
}

##load data and get all reaction IDs
data(H.sapiens_Recon_1)
model <- H.sapiens_Recon_1@model
##get all reaction identifiers
sapply(model@reactions, id)




}

\keyword{hsapiens}
