\name{M.barkeri_iAF692}
\docType{data}
\alias{M.barkeri_iAF692}
\title{ Metabolic reconstruction of M.barkeri from the BiGG database}
\description{The dataset was generated by downloading the SBML file of the
reconstruction (http://bigg.ucsd.edu/bigg/exportSelect.pl) which was subsequently converted
into an object of class \code{SBML} using the \code{rsbml_read} function
from the \code{rsbml} package.}
\usage{data(M.barkeri_iAF692)}

\format{An sbml object of class rsbml}

\details{Note that the files in the BiGG database fail the unit
consistancy check of the \code{rsbml_read} function. To avoid
unit checking when creating \code{SBML} objects, the substance units in the reaction tags
were parsed out from the database SBML files (see example below).}

\source{http://bigg.ucsd.edu/bigg/exportSelect.pl}

\references{
Feist, A.M., Scholten, J.C.M., Palsson, B.O., Brockman, F.J., and Ideker, T., "Modeling methanogenesis with a genome-scale metabolic reconstruction of Methanosarcina barkeri", Molecular Systems Biology, 2(1):msb4100046-E1-E14 (2006)
}

\examples{
\dontrun{
##The dataset was generated as follows:
##SBML_export.xml was downloaded from http://bigg.ucsd.edu/bigg/exportSelect.pl
##and a newline was added at the end of the file
file <- "SBML_export.xml"
string <- paste(readLines(file), collapse="\n")
##Parse out units to avoid validation error
string <- gsub("units=\".+?\"", "", string)
M.barkeri_iAF692 <- rsbml_read(text=string) 
}

##load data and get all reaction IDs
data(M.barkeri_iAF692)
model <- M.barkeri_iAF692@model
##get all reaction identifiers
sapply(model@reactions, id)

}

\keyword{mbarkeri}
