% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{normalise_daily}
\alias{normalise_daily}
\title{Normalize time serie using hourly over daily relative consumption (%) normalization method}
\usage{
normalise_daily(data, method = "relative", localTimeZone)
}
\arguments{
\item{data}{<data.frame> containing serie to normalise}

\item{method}{<string> Normalization method. Supported methods
relative}
}
\value{
daily normalised timeserie
}
\description{
Normalize time serie using hourly over daily relative consumption (%) normalization method
}
