% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{normalise_daily}
\alias{normalise_daily}
\title{Normalise time series by sum of the day}
\usage{
normalise_daily(data, method = "relative", localTimeZone)
}
\arguments{
\item{data}{<data.frame> containing serie to normalise}

\item{method}{<string> Normalisation method. 
Supported methods: relative}
}
\value{
<data.frame> daily normalised time series
}
\description{
Normalise time series using daily relative normalisation method
}
