% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{normalise_dlc}
\alias{normalise_dlc}
\title{Daily load curve normalisation}
\usage{
normalise_dlc(
  data,
  localTimeZone,
  transformation = "relative",
  inputVars = c("loadCurves"),
  nDayParts = 24,
  holidays = c(),
  method = "range01",
  scalingAttr = NULL,
  balanceOutdoorTemperatures = NULL
)
}
\arguments{
\item{data}{<timeserie>}

\item{localTimeZone}{<string> specifying the local time zone related to the
building in analysis. The format of this time zones are defined by the IANA
Time Zone Database (https://www.iana.org/time-zones).}

\item{transformation}{<string> absolute or relative}

\item{inputVars}{<array> of strings. Possible values: loadCurves, daysWeekend,
daysHolidays, daysWeek, dailyTemperature, dailyConsumption, dailyHdd, dailyCdd,
ratioDailyConsumptionHdd, ratioDailyConsumptionCdd}

\item{nDayParts}{<int> number of part days. Clustering considering
parts of the day as "aggregation" of multiple hours. Default value 24 so
each hour is considered a single part of the day}

\item{holidays}{<array> of dates. Holidays dates that are ignored in clustering phase.}

\item{method}{<string> Normalisation methods supported:
- range01. Min-max normalisation method
- znorm. Z-score normalisation method}

\item{scalingAttr}{<data.frame> it includes the scaling attributes for each
variable}

\item{balanceOutdoorTemperatures}{<array> of floats with the balance temperatures 
used in the degree days calculation. Optional.}
}
\value{
normalised load <data.frame>
}
\description{
Normalise time serie applying pre processing transformations
and considering multiple input variables. Ignoring specific dates if required
and supporting multiple normalisation methods
}
