% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{normalise_range}
\alias{normalise_range}
\title{Range normalization method}
\usage{
normalise_range(
  data,
  lower = 0,
  upper = 1,
  lowerThreshold = NULL,
  upperThreshold = NULL
)
}
\arguments{
\item{data}{<serie> containing serie to normalise}

\item{lower}{<float> lower value}

\item{upper}{<float> upper value}

\item{lowerThreshold}{<float> lower threshold value}

\item{upperThreshold}{<float> upper threshold value}
}
\value{
normalised serie
}
\description{
Range normalization method
}
