% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{normalise_range}
\alias{normalise_range}
\title{Normalze time series using min-max range normalisation method}
\usage{
normalise_range(
  data,
  lower = 0,
  upper = 1,
  lowerThreshold = NULL,
  upperThreshold = NULL,
  scalingAttr = NULL
)
}
\arguments{
\item{data}{<data.frame> containing a set of time series to normalise
or an <array> of numerical values to normalise.}

\item{lower}{<float> lower value of the resultant range.}

\item{upper}{<float> upper value of the resultant range.}

\item{lowerThreshold}{<float> lower threshold value.}

\item{upperThreshold}{<float> upper threshold value.}

\item{scalingAttr}{<data.frame> setting min-max threshold for each variable.
Column names: column names specified in data argument; Row names: min, max.
In case of NULL (default value), lowerThreshold or upperThreshold will be used.}
}
\value{
<list> containing two keys: "values" and "scalingAttr". The
former consists on a <data.frame> or an <array> with the normalised values, 
depending the class of data argument. The latter consists on the scaling 
attributes used for normalisation.
}
\description{
Normalze time series using min-max range normalisation method
}
