% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{normalise_zscore}
\alias{normalise_zscore}
\title{Zscore normalization method}
\usage{
normalise_zscore(data)
}
\arguments{
\item{data}{<timeSeries> containing serie to normalise}
}
\value{
normalised timeseries
}
\description{
Zscore normalization method
}
