% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{normalise_zscore}
\alias{normalise_zscore}
\title{Z-Normalisation}
\usage{
normalise_zscore(data, scalingAttr = NULL)
}
\arguments{
\item{data}{<data.frame> containing serie to normalise}

\item{scalingAttr}{<data.frame> setting average and standard deviation 
thresholds for each variable. Column names: column names specified in 
data argument; Row names: mean, sd.}
}
\value{
<list> containing two keys: "values" and "scalingAttr". The
former consists on a <data.frame> or an <array> with the normalised values, 
depending the class of data argument. The latter consists on the scaling 
attributes used for normalisation.
}
\description{
Normalise time series using Zscore normalisation method
}
