% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R, R/zzz.R
\docType{data}
\name{paramiko}
\alias{paramiko}
\title{paramiko}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.

An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
paramiko

paramiko
}
\description{
paramiko

paramiko module
}
\keyword{datasets}
