\name{Recon2}
\docType{data}
\alias{Recon2}
\title{ Human metabolic reconstruction Recon2}
\description{The dataset was generated by downloading the SBML file of the
reconstruction (http://www.ebi.ac.uk/biomodels-main/MODEL1109130000) which was subsequently converted
into an object of class \code{SBML} using the \code{rsbml_read} function
from the \code{rsbml} package.}
\usage{data(Recon2)}
\format{An sbml object of class rsbml}
\source{http://www.ebi.ac.uk/biomodels-main/MODEL1109130000}
\references{
Thiele I, Swainston N, et al.,"A community-driven global reconstruction
of human metabolism", Nature Biotechnology 31, 419-425 (2013), doi:10.1038/nbt.2488
}


\examples{

\dontrun{
##The dataset was generated as follows:
##MODEL1109130000.xml was downloaded from http://www.ebi.ac.uk/biomodels-main/MODEL1109130000
Recon2 <- rsbml_read("MODEL1109130000.xml") 
}

##load data and get all reaction IDs
data(Recon2)
model <- Recon2@model
##get all reaction identifiers
sapply(model@reactions, id)


}


\keyword{recon2}
