% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{resample}
\alias{resample}
\title{Resample time series to new ISO 8601 timestep}
\usage{
resample(data, timestep)
}
\arguments{
\item{data}{<data.frame> describing the input time series to be resampled. 
Time column: 'time', value column: 'value'.}

\item{timeStep}{<string> A string in ISO 8601 format representing the period
or timestep (e.g. "PT15M","PT1H", "P3M", "P1D" ,...).}
}
\value{
<data.frame> corresponding to the resampled time series
}
\description{
Time series is upsampled creating synthetic forward fill time series
in case of higher frequency timestep required. On contrary, is 
downsampled maintaing current values but repeated indexes in case 
of lower frequency timestep required.
}
