% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{round_up}
\alias{round_up}
\title{Round up value or array}
\usage{
round_up(x)
}
\arguments{
\item{x}{<array> of floats to be rounded}
}
\value{
<array> of floats corresponding to the rounded value
}
\description{
This function is used for rounding
}
