\name{S.aureus_iSB619}
\docType{data}
\alias{S.aureus_iSB619}
\title{ Metabolic reconstruction of S.aureus from the BiGG database}
\description{The dataset was generated by downloading the SBML file of the
reconstruction (http://bigg.ucsd.edu/bigg/exportSelect.pl) which was subsequently converted
into an object of class \code{SBML} using the \code{rsbml_read} function
from the \code{rsbml} package.}
\usage{data(S.aureus_iSB619)}
\format{An sbml object of class rsbml}

\details{Note that the files in the BiGG database fail the unit
  consistancy check of the \code{rsbml_read} function. To avoid
  unit checking when creating \code{SBML} objects, the substance units in the reaction tags
  were parsed out from the database SBML files (see example below).}


\source{http://bigg.ucsd.edu/bigg/exportSelect.pl}

\references{
Becker, S.A. and Palsson, B.O., Genome-scale reconstruction of the metabolic network in Staphylococcus aureus N315: an initial draft to the two-dimensional annotation, BMC Microbiology, 5(1):8 (2005)
}

\examples{

\dontrun{
##The dataset was generated as follows:
##SBML_export.xml was downloaded from http://bigg.ucsd.edu/bigg/exportSelect.pl
##and a newline was added at the end of the file
file <- "SBML_export.xml"
string <- paste(readLines(file), collapse="\n")
##Parse out units to avoid validation error
string <- gsub("units=\".+?\"", "", string)
S.aureus_iSB619 <- rsbml_read(text=string) 
}

##load data and get all reaction IDs
data(S.aureus_iSB619)
model <- S.aureus_iSB619@model
##get all reaction identifiers
sapply(model@reactions, id)


}

\keyword{saureus}
