\name{S.cerevisiae_iND750}
\docType{data}
\alias{S.cerevisiae_iND750}
\title{ Metabolic reconstruction of S.cerevisiae from the BiGG database}
\description{The dataset was generated by downloading the SBML file of the
reconstruction (http://bigg.ucsd.edu/bigg/exportSelect.pl) which was subsequently converted
into an object of class \code{SBML} using the \code{rsbml_read} function
from the \code{rsbml} package.}
\usage{data(S.cerevisiae_iND750)}
\format{An sbml object of class rsbml}

\details{Note that the files in the BiGG database fail the unit
consistancy check of the \code{rsbml_read} function. To avoid
unit checking when creating \code{SBML} objects, the substance units in the reaction tags
were parsed out from the database SBML files (see example below).}

\source{http://bigg.ucsd.edu/bigg/exportSelect.pl}

\references{
Duarte, N.C., Herrgard, M.J., and Palsson, B.O., " Reconstruction and Validation of Saccharomyces cerevisiae iND750, a Fully Compartmentalized Genome-scale Metabolic Model", Genome Research, 14: 1298-1309 (2004)
}

\examples{

\dontrun{
##The dataset was generated as follows:
##SBML_export.xml was downloaded from http://bigg.ucsd.edu/bigg/exportSelect.pl
##and a newline was added at the end of the file
file <- "SBML_export.xml"
string <- paste(readLines(file), collapse="\n")
##Parse out units to avoid validation error
string <- gsub("units=\".+?\"", "", string)
S.cerevisiae_iND750 <- rsbml_read(text=string) 
}

##load data and get all reaction IDs
data(S.cerevisiae_iND750)
model <- S.cerevisiae_iND750@model
##get all reaction identifiers
sapply(model@reactions, id)


}

\keyword{scervisiae}
