% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_group_create.R
\name{security_group_create}
\alias{security_group_create}
\title{security_group_create}
\usage{
security_group_create(group_name = NA, description = "Add a description")
}
\arguments{
\item{group_name}{Name of security group}

\item{description}{}
}
\description{
security_group_create
}
